% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_present}
\alias{all_present}
\title{Are all accounts present in debit and credit columns}
\usage{
all_present(pos, neg)
}
\description{
Used in \code{deb_account_summary()}, \code{deb_credit()}, and \code{deb_debit()}
if statement to see whether all accounts have both a credit and debit
transaction or not.
}
\keyword{internal}
