% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{arithmetic}
\alias{arithmetic}
\alias{vec_arith.deb_lsd}
\alias{vec_arith.deb_lsd.default}
\alias{vec_arith.deb_lsd.deb_lsd}
\alias{vec_arith.deb_lsd.numeric}
\alias{vec_arith.numeric.deb_lsd}
\alias{vec_arith.deb_lsd.MISSING}
\alias{vec_arith.deb_decimal}
\alias{vec_arith.deb_decimal.default}
\alias{vec_arith.deb_decimal.deb_decimal}
\alias{vec_arith.deb_decimal.numeric}
\alias{vec_arith.numeric.deb_decimal}
\alias{vec_arith.deb_decimal.MISSING}
\alias{vec_arith.deb_tetra}
\alias{vec_arith.deb_tetra.default}
\alias{vec_arith.deb_tetra.deb_tetra}
\alias{vec_arith.deb_tetra.numeric}
\alias{vec_arith.numeric.deb_tetra}
\alias{vec_arith.deb_tetra.MISSING}
\alias{vec_arith.deb_lsd.deb_decimal}
\alias{vec_arith.deb_decimal.deb_lsd}
\alias{vec_arith.deb_lsd.deb_tetra}
\alias{vec_arith.deb_tetra.deb_lsd}
\alias{vec_arith.deb_tetra.deb_decimal}
\alias{vec_arith.deb_decimal.deb_tetra}
\title{Arithmetic operations for debkeepr}
\usage{
\method{vec_arith}{deb_lsd}(op, x, y, ...)

\method{vec_arith.deb_lsd}{default}(op, x, y, ...)

\method{vec_arith.deb_lsd}{deb_lsd}(op, x, y, ...)

\method{vec_arith.deb_lsd}{numeric}(op, x, y, ...)

\method{vec_arith.numeric}{deb_lsd}(op, x, y, ...)

\method{vec_arith.deb_lsd}{MISSING}(op, x, y, ...)

\method{vec_arith}{deb_decimal}(op, x, y, ...)

\method{vec_arith.deb_decimal}{default}(op, x, y, ...)

\method{vec_arith.deb_decimal}{deb_decimal}(op, x, y, ...)

\method{vec_arith.deb_decimal}{numeric}(op, x, y, ...)

\method{vec_arith.numeric}{deb_decimal}(op, x, y, ...)

\method{vec_arith.deb_decimal}{MISSING}(op, x, y, ...)

\method{vec_arith}{deb_tetra}(op, x, y, ...)

\method{vec_arith.deb_tetra}{default}(op, x, y, ...)

\method{vec_arith.deb_tetra}{deb_tetra}(op, x, y, ...)

\method{vec_arith.deb_tetra}{numeric}(op, x, y, ...)

\method{vec_arith.numeric}{deb_tetra}(op, x, y, ...)

\method{vec_arith.deb_tetra}{MISSING}(op, x, y, ...)

\method{vec_arith.deb_lsd}{deb_decimal}(op, x, y, ...)

\method{vec_arith.deb_decimal}{deb_lsd}(op, x, y, ...)

\method{vec_arith.deb_lsd}{deb_tetra}(op, x, y, ...)

\method{vec_arith.deb_tetra}{deb_lsd}(op, x, y, ...)

\method{vec_arith.deb_tetra}{deb_decimal}(op, x, y, ...)

\method{vec_arith.deb_decimal}{deb_tetra}(op, x, y, ...)
}
\arguments{
\item{op}{Arithmetic operation.}

\item{x, y}{Vectors.}

\item{...}{For future expansion}
}
\value{
A \code{deb_lsd}, \code{deb_tetra}, \code{deb_decimal} or \code{numeric} vector
depending on the inputs and arithmetic operator.
}
\description{
Arithmetic operations for debkeepr
}
