% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{comparison}
\alias{comparison}
\alias{vec_proxy_equal.deb_lsd}
\alias{vec_proxy_compare.deb_lsd}
\alias{vec_proxy_equal.deb_tetra}
\alias{vec_proxy_compare.deb_tetra}
\title{Equality and comparison}
\usage{
\method{vec_proxy_equal}{deb_lsd}(x, ...)

\method{vec_proxy_compare}{deb_lsd}(x, ...)

\method{vec_proxy_equal}{deb_tetra}(x, ...)

\method{vec_proxy_compare}{deb_tetra}(x, ...)
}
\arguments{
\item{x}{A deb_lsd vector.}

\item{...}{Arguments passed on to further methods.}
}
\value{
A data frame or numeric vector to be used for comparison.
}
\description{
Equality and comparison
}
