% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-accounts.R
\name{deb_account_summary}
\alias{deb_account_summary}
\title{Calculate credit, debit, and current values of accounts}
\usage{
deb_account_summary(df, credit = credit, debit = debit, lsd = lsd,
  round = 5, na.rm = FALSE)
}
\arguments{
\item{df}{A data frame containing transactions between accounts and values
in the form of an lsd list column.}

\item{credit}{Credit column: Unquoted name of the credit variable. This
represents the accounts that discharge the transactional values or from
which the values derive. Default is credit. The \code{credit} column must be
of the same class as the \code{debit} column.}

\item{debit}{Debit column: Unquoted name of the debit variable. This
represents the accounts that receive the transactional values. Default is
debit. The \code{debit} column must be of the same class as the \code{credit}
column.}

\item{lsd}{lsd list column: Unquoted name of an lsd list column of pounds,
shillings, and pence values with a bases attribute.}

\item{round}{Round pence to specified number of decimal places. Default
is 5. Rounding of the pence value takes place throughout the function and
is not simply done at the end of the function call. This means that the
credit, debit, and current values will add together correctly. However,
the rounded values may be different than running the function without
rounding and rounding pence values at the end.}

\item{na.rm}{Logical. Passed on to \code{na.rm} argument in \code{\link[=sum]{sum()}}. Whether
missing values (NA) should be removed. Default is \code{FALSE}.}
}
\value{
Returns a tibble with three rows for each account present in the
credit and/or debit variables of \code{df} with the total credit, debit, and
current values of the accounts in an lsd list column. If an account does
not have any credit or debit transactions, it will not have a summary row
for that type of relation.
}
\description{
Calculate the total credit, debit, and the current values for all accounts
in a transaction data frame with an lsd list column. Credits and debits are
both returned as positive numbers. If an account has more credit than debit,
the current value will be returned as positive values. If the debit is
greater, the current value will be returned as negative values.
}
\details{
\code{deb_account_summary()} is similar to \code{\link[=deb_account]{deb_account()}}, but it returns the
values for all accounts in a transaction data frame instead of one. If you
only want to see the current values for single accounts, see
\code{\link[=deb_current]{deb_current()}}.

\code{deb_account_summary()} is part of a family of functions meant to be used on
data frames that contain transactions between accounts in an account
book. The data frame should possess a similar structure to a
\href{https://www.jessesadler.com/post/network-analysis-with-r/#nodes-edges}{directed edge list}.
In this context, credit and debit variables contain the ids of the accounts
and act as the edges of the network. Following the
\href{https://en.wikipedia.org/wiki/Debits_and_credits}{nomenclature of accounting},
the credit variable represents the account that the transactional value is
from, while the debit variable represents the account that receives the
value. Thus, from the credit account to the debit account.
}
\examples{
library(tibble)

# Create a transactions data frame
lsd_list <- deb_lsd(l = c(10, 6, 4, 7, 9),
                    s = c(15, 3, 11, 11, 2),
                    d = c(6, 11, 7, 8, 11),
                    bases = c(20, 12))
trans <- tibble(credit = c("a", "b", "b", "a", "c"),
                debit = c("b", "a", "a", "c", "a"),
                lsd = lsd_list)

# Credit, debit, and current values of accounts present in trans
deb_account_summary(df = trans,
                    credit = credit,
                    debit = debit,
                    lsd = lsd)

}
\seealso{
Other lsd account functions: \code{\link{deb_account}},
  \code{\link{deb_balance}}, \code{\link{deb_credit}},
  \code{\link{deb_current}}, \code{\link{deb_debit}},
  \code{\link{deb_open}}
}
\concept{lsd account functions}
