% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-arithmetic.R
\name{deb_add_mutate}
\alias{deb_add_mutate}
\title{Addition of pounds, shillings, and pence in a data frame}
\usage{
deb_add_mutate(df, l = l, s = s, d = d, lsd, bases = c(20, 12),
  replace = FALSE, suffix = ".1")
}
\arguments{
\item{df}{A data frame that contains pounds, shillings, and pence variables.}

\item{l}{Pounds column: Unquoted name of a numeric variable corresponding
to pounds. Default is l.}

\item{s}{Shillings column: Unquoted name of numeric variable corresponding
to shillings. Default is s.}

\item{d}{Pence column: Unquoted name of numeric variable corresponding to
pence. Default is d.}

\item{lsd}{Numeric vector of length 3 that represents the lsd value to be
added to the \code{l}, \code{s}, and \code{d} variables. The first position of the vector
represents the pounds value or l. The second position represents the
shillings value or s. And the third position represents the pence value
or d.}

\item{bases}{Numeric vector of length 2 used to specify the bases for
the s or solidus and d or denarius values in the lsd variables. Default is
\code{c(20, 12)}, which conforms to the most widely used system of 1 libra =
20 solidi and 1 solidus = 12 denarii. This argument makes it possible to
use alternative bases for the solidus and denarius values that were also
in use.}

\item{replace}{Logical (default \code{FALSE}): when \code{TRUE} the new pounds,
shillings, and pence variables will replace the original ones.}

\item{suffix}{Suffix added to the column names for the pounds, shillings, and
pence columns representing the added values so that they are
distinguished from the original pounds, shillings, and pence columns.
Default is ".1". Should be a character vector of length 1.}
}
\value{
Returns a data frame with three new variables of pounds, shillings,
and pence representing the added values.
}
\description{
Uses \code{\link[dplyr:mutate]{dplyr::mutate()}} to add pounds, shillings, and pence by a given
pounds, shillings, and pence value. The values are returned in the form of
three new variables representing the calculated pounds, shillings and pence.
}
\details{
See \code{\link[=deb_sum_df]{deb_sum_df()}} to get a sum of the lsd values in a data frame.
}
\examples{
# Add a data frame of lsd values by £5 15s. 8d
example <- data.frame(l = c(35, 10, 26, 12),
                      s = c(10, 18, 11, 16),
                      d = c(9, 11, 10, 5))
deb_add_mutate(example, l, s, d, lsd = c(5, 15, 8))

# Replace the existing pounds, shillings, and pence
# with added values
deb_add_mutate(example, l, s, d,
               lsd = c(5, 15, 8),
               replace = TRUE)

}
\seealso{
Other lsd arithmetic functions: \code{\link{deb_add}},
  \code{\link{deb_divide_mutate}},
  \code{\link{deb_divide}},
  \code{\link{deb_multiply_mutate}},
  \code{\link{deb_multiply}},
  \code{\link{deb_subtract_mutate}},
  \code{\link{deb_subtract}}, \code{\link{deb_sum_df}},
  \code{\link{deb_sum}}
}
