% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casting.R
\name{deb_as_list}
\alias{deb_as_list}
\title{Cast \code{deb_lsd} to a list of lsd values}
\usage{
deb_as_list(x, ...)
}
\arguments{
\item{x}{A \code{deb_lsd} vector to cast to a list of lsd values.}

\item{...}{Arguments passed on to further methods.}
}
\value{
A list of numeric vectors of length 3, corresponding to lsd values.
}
\description{
Cast a \code{deb_lsd} vector to a list of numeric vectors containing lsd values.
}
\details{
\code{deb_as_list()} turns a \code{deb_lsd} vector into a list of numeric vectors of
length 3. It is the inverse of \code{deb_as_lsd()}. Compare to \code{as.list()},
which creates a list of \code{deb_lsd} vectors or \code{unclass()}, which creates a
list of length 3 with numeric vectors for pounds, shillings, and pence.
}
\examples{

# deb_lsd vector
x <- deb_lsd(l = 0:3, s = 4:7, d = 8:11)

deb_as_list(x)

# This is the inverse of `deb_as_lsd()` of a list of lsd values
y <- deb_as_list(x)

identical(x, deb_as_lsd(y))

}
\seealso{
\code{\link[=deb_as_lsd]{deb_as_lsd()}} for the inverse of \code{deb_as_list()}.
}
