% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-conversion.R
\name{deb_convert_bases_mutate}
\alias{deb_convert_bases_mutate}
\title{Convert between different bases for shillings and pence units in a data
frame}
\usage{
deb_convert_bases_mutate(df, l = l, s = s, d = d, bases1, bases2,
  ratio = 1, round = 5, suffix = ".1")
}
\arguments{
\item{df}{A data frame that contains pounds, shillings, and pence variables.}

\item{l}{Pounds column: Unquoted name of a numeric variable corresponding
to the pounds or librae unit. Default is l.}

\item{s}{Shillings column: Unquoted name of numeric variable corresponding
to the shillings or solidi unit. Default is s.}

\item{d}{Pence column: Unquoted name of numeric variable corresponding to
the pence or denarii unit. Default is d.}

\item{bases1}{Numeric vector of length 2 used to specify the bases for the
shillings or s and pence or d values of \code{l}, \code{s}, and \code{d} variables.}

\item{bases2}{Numeric vector of length 2 used to specify the bases for the
shillings or s and pence or d values to which \code{l}, \code{s}, and \code{d} will be
converted.}

\item{ratio}{The ratio between the two currencies that possess different
bases. This is the value by which \code{l}, \code{s}, and \code{d} variables are
multiplied. Numeric vector of length 1 with the default of \code{1}.}

\item{round}{Round pence to specified number of decimal places.
Default is 5. Set to 0 to return pence as whole numbers.}

\item{suffix}{Suffix added to the column names for the pounds, shillings,
and pence columns representing the converted values to distinguish them
from the original pounds, shillings, and pence columns. Default is ".1".
Must be a character vector of length 1.}
}
\value{
Returns a data frame with three new variables of pounds, shillings,
and pence corresponding to the shillings and pence bases in \code{bases2}.
}
\description{
Uses \code{\link[dplyr:mutate]{dplyr::mutate()}} to convert pounds, shillings, and pence currencies
that use different bases for the shillings and pence variables in a data
frame. The converted values are returned in the form of three new variables
representing the calculated pounds, shillings and pence for the new bases.
}
\details{
\code{deb_convert_bases_mutate()} is similar to \code{\link[=deb_exchange_mutate]{deb_exchange_mutate()}}, but the
latter performs exchanges on currencies that share the same shillings and
pence bases.
}
\examples{
# Conversion between pounds Flemish of 20 shillings and 12 pence
# to guilders of 20 stuivers and 16 penningen at the rate of
# 6 guilders equals £1 Flemish
flemish <- data.frame(l = c(204, 82, 36),
                      s = c(3, 15, 12),
                      d = c(3, 9, 6))
deb_convert_bases_mutate(df = flemish,
                         l = l, s = s, d = d,
                         bases1 = c(20, 12),
                         bases2 = c(20, 16),
                         ratio = 6)

# Convert from guilders to pounds Flemish
# Flip the bases argument and change the ratio
guilders <- data.frame(l = c(1224, 101, 225),
                       s = c(19, 5, 13),
                       d = c(8, 13, 15))
deb_convert_bases_mutate(df = guilders,
                         l = l, s = s, d = d,
                         bases1 = c(20, 16),
                         bases2 = c(20, 12),
                         ratio = 1/6)

}
