% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-decimalization-df.R
\name{deb_d_lsd_mutate}
\alias{deb_d_lsd_mutate}
\title{Conversion of decimalized pence into pounds, shillings, and pence variables}
\usage{
deb_d_lsd_mutate(df, pence, l_column = l, s_column = s, d_column = d,
  bases = c(20, 12), round = 5, suffix = ".1")
}
\arguments{
\item{df}{A data frame that contains a column of pence.}

\item{pence}{Decimalized pence column: Unquoted name of a numeric variable
corresponding to pence The variable that will be mutated into pounds,
shillings, and pence variables.}

\item{l_column}{An unquoted name for the pounds column created by the
function. Default is l.}

\item{s_column}{An unquoted name for the shillings column created by
the function. Default is s.}

\item{d_column}{An unquoted name for the pence column created by the
function. Default is d.}

\item{bases}{Numeric vector of length 2 used to specify the bases for the
shillings or s and pence or d units. Default is \code{c(20, 12)}, which
conforms to the most widely used system of 1 pound = 20 shillings and
1 shilling = 12 pence. If \code{lsd} is of class lsd, the bases attribute will
be used in the place of this argument.}

\item{round}{Round pence unit to specified number of decimal places.
Default is 5. Set to 0 to return pence as whole number.}

\item{suffix}{Suffix added to \code{l_column}, \code{s_column}, or \code{d_column} if \code{df}
already contains variables with the same names. Default is ".1". Must be a
character vector of length 1.}
}
\value{
Returns a data frame with three new variables of pounds, shillings,
and pence.
}
\description{
Uses \code{\link[dplyr:mutate]{dplyr::mutate()}} to add equivalent pounds, shillings, and pence
variables to a data frame that contains a pence variable. Thus, the
function converts pence into the lsd system of pounds, shillings,
and pence.
}
\examples{
# Create equivalent pounds, shillings, and pence
# variables from a pence variable
example <- data.frame(pence = c(1998, -1998, 387, -5378))
deb_d_lsd_mutate(df = example, pence = pence)

# You can choose the names for the created columns
example \%>\%
  deb_d_lsd_mutate(pence = pence,
                   l_column = librae,
                   s_column = solidi,
                   d_column = pence)

}
