% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-mutate.R
\name{deb_d_mutate}
\alias{deb_d_mutate}
\title{Mutate decimal pence into pounds, shillings, and pence variables}
\usage{
deb_d_mutate(df, denarii, l_column = l, s_column = s, d_column = d,
  round = 3, suffix = ".1")
}
\arguments{
\item{df}{A data frame that contains a column of pence.}

\item{denarii}{Pence column: Unquoted name of a numeric variable
corresponding to pence. This is the variable that will be
mutated into pounds, shillings, and pence variables.}

\item{l_column}{An unquoted name for the pounds column created by the
function. Default is l.}

\item{s_column}{An unquoted name for the shillings column created by
the function. Default is s.}

\item{d_column}{An unquoted name for the pence column created by the
function. Default is d.}

\item{round}{round pence to specified number of decimal places.
Default is 3. Set to 0 to return pence as whole numbers.}

\item{suffix}{If the data frame already contains variables with the same
names as \code{l_column}, \code{s_column}, or \code{d_column}, this suffix will be
added to the new variables to distinguish them. Default is ".1".}
}
\value{
Returns a data frame with three new variables of pounds, shillings,
and pence.
}
\description{
Uses \code{\link[dplyr:mutate]{dplyr::mutate()}} to add equivalent pounds, shillings, and pence
variables to a data frame that contains a pence variable. Thus, the
function converts pence into the lsd system of pounds, shillings,
and pence.
}
\examples{
# Create equivalent pounds, shillings, and pence
# variables from a pence variable
example <- data.frame(pence = c(1998, -1998, 387, -5378))

deb_d_mutate(example, pence)

# You can choose the names for the created columns
example \%>\%
  deb_d_mutate(pence,
               l_column = librae,
               s_column = solidi,
               d_column = denarii)

}
