% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-accounts.R
\name{deb_debit}
\alias{deb_debit}
\title{Calculate the total debit of accounts}
\usage{
deb_debit(df, credit = credit, debit = debit, lsd = lsd, round = 5,
  na.rm = FALSE)
}
\arguments{
\item{df}{A data frame containing transactions between accounts and values
in the form of an lsd list column.}

\item{credit}{Credit column: Unquoted name of the credit variable. This
represents the accounts that discharge the transactional values or from
which the values derive. Default is credit. The \code{credit} column must be
of the same class as the \code{debit} column.}

\item{debit}{Debit column: Unquoted name of the debit variable. This
represents the accounts that receive the transactional values. Default is
debit. The \code{debit} column must be of the same class as the \code{credit}
column.}

\item{lsd}{lsd list column: Unquoted name of an lsd list column of pounds,
shillings, and pence values with a bases attribute.}

\item{round}{Round pence to specified number of decimal places. Default
is 5. Rounding to whole numbers may give results that do not add together
due to rounding rules.}
}
\value{
Returns a tibble with one row for each account in \code{df} with the
values in an lsd list column. The values represent the total value
received by each account from other accounts within \code{df}.
}
\description{
Calculate the total debit of accounts in a transactions data frame with an
lsd list column.
}
\details{
\code{deb_debit()} is similar to \code{\link[=deb_account_summary]{deb_account_summary()}}, but it only returns
the debit values for the accounts in \code{df}. Accounts that are not debtor in
any transactions appear with a value of £0. See \code{\link[=deb_credit]{deb_credit()}} to return the
credit totals for the accounts in \code{df}.

\code{deb_debit()} is part of a family of functions meant to be used on
data frames that contain transactions between accounts in an account
book. The data frame should possess a similar structure to a
\href{https://www.jessesadler.com/post/network-analysis-with-r/#nodes-edges}{directed edge list}.
In this context, credit and debit variables contain the ids of the accounts
and act as the edges of the network. Following the
\href{https://en.wikipedia.org/wiki/Debits_and_credits}{nomenclature of accounting},
the credit variable represents the account that the transactional value is
from, while the debit variable represents the account that receives the
value. Thus, from the credit account to the debit account.
}
\examples{
library(tibble)

# Create a transactions data frame
lsd_list <- deb_lsd(l = c(10, 6, 4, 7, 9),
                    s = c(15, 3, 11, 11, 2),
                    d = c(6, 11, 7, 8, 11),
                    bases = c(20, 12))
trans <- tibble(credit = c("a", "b", "b", "a", "c"),
                debit = c("b", "a", "a", "c", "a"),
                lsd = lsd_list)

# Total debit of accounts present in trans
deb_debit(df = trans,
          credit = credit,
          debit = debit,
          lsd = lsd)

}
\seealso{
Other lsd account functions: \code{\link{deb_account_summary}},
  \code{\link{deb_account}}, \code{\link{deb_balance}},
  \code{\link{deb_credit}}, \code{\link{deb_current}},
  \code{\link{deb_open}}
}
\concept{lsd account functions}
