% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal-class.R
\name{deb_decimal}
\alias{deb_decimal}
\title{A decimalized class for pounds, shillings, and pence values}
\usage{
deb_decimal(x = double(), unit = c("l", "s", "d"), bases = c(20, 12))
}
\arguments{
\item{x}{A numeric vector representing the decimalized values of a
non-decimal currency.}

\item{unit}{A character vector of length one indicating the unit for the
decimalized values, either \code{"l"} (libra, the default), \code{"s"} (solidus),
or \code{"d"} (denarius).}

\item{bases}{Numeric vector of length 2 used to specify the bases for the
solidus or s and denarius or d units. Default is \code{c(20, 12)}, which
conforms to the most widely used system of 1 pound = 20 shillings and
1 shilling = 12 pence.}
}
\value{
Returns a vector of class \code{deb_decimal}.
}
\description{
Create a vector of class \code{deb_decimal} to integrate non-decimal currencies
into standardized forms of analysis provided by R.
}
\details{
The \code{deb_decimal} class and the \code{debkeepr} package use the nomenclature of
\href{https://en.wikipedia.org/wiki/£sd}{l, s, and d} to represent pounds,
shillings, and pence units. The abbreviations derive from the Latin terms
\href{https://en.wikipedia.org/wiki/French_livre}{libra},
\href{https://en.wikipedia.org/wiki/Solidus_(coin)}{solidus}, and
\href{https://en.wikipedia.org/wiki/Denarius}{denarius}. In the 8th century a
solidus came to represent 12 denarii coins, and, for a time at least, 240
denarii were made from one libra or pound of silver. The custom of counting
coins in dozens (solidi) and scores of dozens (librae) spread throughout the
Carolingian Empire and became engrained in much of Europe. However, a
variety of accounting systems arose at different times that used
\href{https://en.wikipedia.org/wiki/Non-decimal_currency}{other bases} for the
solidus and denarius units. The \code{bases} attribute of \code{deb_decimal} vectors
makes it possible to specify alternative bases for the solidus and denarius
units, while the \code{unit} attribute identifies the decimalized unit: either
libra, solidus, or denarius.
}
\examples{

deb_decimal(c(5.25, 3.825, 8.5))

# Set the unit of the deb_decimal vector
deb_decimal(c(105, 76.5, 170), unit = "s")
deb_decimal(c(1260, 918, 240), unit = "d")

# Set the bases of the deb_decimal vector
deb_decimal(c(5.25, 3.825, 8.5), bases = c(60, 16))

# Create a prototype or vector of length 0
deb_decimal()
}
\seealso{
The \code{deb_decimal} class works in concert with the \code{deb_lsd} class, which
maintains the tripartite unit structure of most non-decimal currencies.
See \code{\link[=deb_lsd]{deb_lsd()}}.
}
