% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{deb_df_to_list}
\alias{deb_df_to_list}
\title{Data frame of pounds, shillings, and pence to an lsd object}
\usage{
deb_df_to_list(df, l = l, s = s, d = d, bases = c(20, 12))
}
\arguments{
\item{df}{A data frame that contains pounds, shillings, and pence variables.}

\item{l}{Pounds column: Unquoted name of a numeric variable corresponding
to the pounds or librae unit. Default is l.}

\item{s}{Shillings column: Unquoted name of numeric variable corresponding
to the shillings or solidi unit. Default is s.}

\item{d}{Pence column: Unquoted name of numeric variable corresponding to
the pence or denarii unit. Default is d.}

\item{bases}{Numeric vector of length 2 used to specify the bases for the
shillings or s and pence or d units. Default is \code{c(20, 12)}, which
conforms to the most widely used system of 1 pound = 20 shillings and
1 shilling = 12 pence. If \code{lsd} is of class lsd, the bases attribute will
be used in the place of this argument.}
}
\value{
Returns an lsd object with a bases attribute. All variables in \code{df}
aside from \code{l}, \code{s}, and \code{d} will be dropped.
}
\description{
Transform a data frame that has pounds, shillings, and pence variables into
an lsd object.
}
\details{
\code{df_to_lsd_list} is a helper function to make it easier to convert between
the two primary types of input objects of multiple lsd values used by
\code{debkeepr}. See \code{\link[=deb_list_to_df]{deb_list_to_df()}} to convert a list of lsd values to a
data frame with pounds, shillings, and pence variables.
}
\examples{
# Convert a data frame with pounds, shillings, and pence variables
# into a list an lsd object
example <- data.frame(l = c(4, -9, 25),
                      s = c(14, -5, 15),
                      d = c(9, -1, 6))
deb_df_to_list(example, l = l, s = s, d = d)

# All variables aside from l, s, and d will be dropped
example2 <- data.frame(credit = c("a", "b", "a", "c"),
                       debit = c("b", "a", "c", "a"),
                       l = c(10, 10, 7, 9),
                       s = c(15, 15, 11, 2),
                       d = c(6, 6, 8, 11))
deb_df_to_list(example2, l = l, s = s, d = d)

}
\seealso{
Other helper functions: \code{\link{deb_list_to_df}}
}
\concept{helper functions}
