% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-arithmetic.R
\name{deb_divide_mutate}
\alias{deb_divide_mutate}
\title{Division of pounds, shillings, and pence in a data frame}
\usage{
deb_divide_mutate(df, l = l, s = s, d = d, x, bases = c(20, 12),
  round = 5, replace = FALSE, suffix = ".1")
}
\arguments{
\item{df}{A data frame that contains pounds, shillings, and pence variables.}

\item{l}{Pounds column: Unquoted name of a numeric variable corresponding
to the pounds or librae unit. Default is l.}

\item{s}{Shillings column: Unquoted name of numeric variable corresponding
to the shillings or solidi unit. Default is s.}

\item{d}{Pence column: Unquoted name of numeric variable corresponding to
the pence or denarii unit. Default is d.}

\item{x}{Divisor. The value by which \code{lsd} will be divided. A numeric vector
of length 1.}

\item{bases}{Numeric vector of length 2 used to specify the bases for
the shillings or s and pence or d units in \code{lsd}. Default is \code{c(20, 12)},
which conforms to the most widely used system of 1 pound = 20 shillings
and 1 shilling = 12 pence. This argument makes it possible to use
alternate bases for the shillings (s) and pence (d) units.}

\item{round}{Round pence to specified number of decimal places.
Default is 5. Set to 0 to return pence as whole numbers.}

\item{replace}{Logical (default \code{FALSE}): when \code{TRUE} the new pounds,
shillings, and pence variables will replace the original ones.}

\item{suffix}{Suffix added to the column names for the pounds, shillings,
and pence columns representing the divided values to distinguish them
from the original pounds, shillings, and pence columns. Default is ".1".
Must be a character vector of length 1.}
}
\value{
Returns a data frame with three new variables of pounds, shillings,
and pence representing the divided values.
}
\description{
Uses \code{\link[dplyr:mutate]{dplyr::mutate()}} to divide pounds, shillings, and pence variables by a
divisor (\code{x}). The values are returned in the form of three new variables
representing the calculated pounds, shillings and pence.
}
\examples{
# Divide a data frame of values by 5
example <- data.frame(l = c(356, 10, 26, 12),
                      s = c(100, 18, 11, 16),
                      d = c(98, 11, 10, 5))
deb_divide_mutate(df = example, l = l, s = s, d = d, x = 5)

# Change the suffix added to the new variables
deb_divide_mutate(df = example,
                  l = l, s = s, d = d,
                  x = 5,
                  suffix = ".div5")

# Replace the existing pounds, shillings, and pence
# with divided values
deb_divide_mutate(df = example,
                  l = l, s = s, d = d,
                  x = 5,
                  replace = TRUE)

}
\seealso{
Other lsd arithmetic functions: \code{\link{deb_add_mutate}},
  \code{\link{deb_add}}, \code{\link{deb_divide}},
  \code{\link{deb_multiply_mutate}},
  \code{\link{deb_multiply}},
  \code{\link{deb_subtract_mutate}},
  \code{\link{deb_subtract}}, \code{\link{deb_sum_df}},
  \code{\link{deb_sum}}
}
\concept{lsd arithmetic functions}
