% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-interest.R
\name{deb_interest}
\alias{deb_interest}
\title{Interest of pounds, shillings, and pence values}
\usage{
deb_interest(lsd, interest, duration = 1, with_principal = TRUE,
  bases = c(20, 12), round = 5)
}
\arguments{
\item{interest}{Interest rate in decimal form. A numeric vector of length 1.}

\item{duration}{Duration over which the interest is calculated, or the
number of times that the interest should be added. A numeric vector of
length 1. Default is 1.}

\item{with_principal}{Logical (default \code{TRUE}): when \code{TRUE} the resulting
pounds, shillings and pence value(s) will include the principal as well as
the interest. When \code{FALSE} only the interest will be returned.}
}
\value{
Returns an object of class lsd with a bases attribute.
}
\description{
Calculate the interest of pounds, shillings, and pence given an interest
rate and a duration. The function does not calculate compound interest.
}
\examples{
# Calculate the interest with the principal over a certain duration
# If £10.14.5 were lent over a period of 5 years at 6.25\%
deb_interest(lsd = c(10, 14, 5), interest = 0.0625, duration = 5)

# If £10.14.5 were lent over a period of 5 years at 8\%
deb_interest(lsd = c(10, 14, 5), interest = 0.08, duration = 5)

# Or you can calculate only the interest
deb_interest(lsd = c(10, 14, 5),
             interest = 0.0625,
             duration = 5,
             with_principal = FALSE)

# Interest for an lsd value with alternative bases
deb_interest(lsd = c(10, 14, 5),
             interest = 0.0625,
             duration = 5,
             bases = c(20, 16))

# Interest of an object of class lsd will use the bases attribute
lsd <- deb_as_lsd(lsd = c(10, 14, 5), bases = c(20, 16))
deb_interest(lsd = lsd, interest = 0.0625, duration = 5)

# Interest of a list of lsd values at a single rate
lsd_list <- list(c(40, 5, 9), c(29, 7, 1), c(35, 6, 5))
deb_interest(lsd = lsd_list, interest = 0.08, duration = 5)

# Or an lsd object with alternative bases
lsd_list2 <- deb_as_lsd(lsd = lsd_list, bases = c(20, 16))
deb_interest(lsd = lsd_list2, interest = 0.08, duration = 5)

}
