% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-interest.R
\name{deb_interest_mutate}
\alias{deb_interest_mutate}
\title{Calculate the interest of pounds, shillings, and pence}
\usage{
deb_interest_mutate(df, l = l, s = s, d = d, interest = 0.0625,
  duration = 1, with_principal = TRUE, suffix = ".interest", round = 3)
}
\arguments{
\item{df}{A data frame that contains columns with pounds, shillings,
and pence variables.}

\item{l}{Pounds column: Unquoted name of a numeric variable corresponding
to pounds. Default is l.}

\item{s}{Shillings column: Unquoted name of numeric variable corresponding
to shillings. Default is s.}

\item{d}{Pence column: Unquoted name of numeric variable corresponding to
pence. Default is d.}

\item{interest}{Interest rate in decimal form. A numeric vector
of length 1. Default is 0.0625, or 6.25 percent, which was default
interest rate in the Low Countries at the end of the sixteenth century.}

\item{duration}{Duration over which the interest is calculated, or
number of times that the interest should be added. A numeric vector
of length 1. Default is 1.}

\item{with_principal}{Logical (default \code{TRUE}: when \code{TRUE} the resulting
pounds, shillings and pence will include the interest charged as well
as the principal. When \code{FALSE} only the interest will be shown.}

\item{suffix}{Suffix added to the column names for the pounds,
shillings, and pence columns representing the interest so that
they are distinguished from the pounds, shillings, and pence
columns of the principal used to make the calculation. Default is
".interest". Should be a character vector of length 1.}

\item{round}{round pence to specified number of decimal places.
Default is 3. Set to 0 if you want pence to always be a whole number.}
}
\value{
Returns a data frame with three new variables of pounds, shillings,
and pence representing the calculated interest.
}
\description{
Uses \code{\link[dplyr:mutate]{dplyr::mutate()}} to calculate the interest of pounds,
shillings, and pence variables given an interest rate and a duration.
The interest—with or without the principal included—is returned in the
form of three new variables representing the calculated pounds, shillings
and pence for the interest. The function does not calculate
compound interest.
}
\examples{
# Calculate the interest of a set of values in a data frame
example <- tibble::tibble(l = c(3, 10, 26, 12),
                          s = c(10, 18, 11, 16),
                          d = c(9, 11, 10, 5))
deb_interest_mutate(example, l, s, d, duration = 5)

# Calculate the interest without the principal
example \%>\%
  deb_interest_mutate(l, s, d,
                      duration = 5,
                      with_principal = FALSE)

# Change the suffix for the new columns from the default
# and round the pence to be whole numbers.
deb_interest_mutate(example,
                    l, s, d,
                    duration = 5,
                    suffix = "_5years",
                    round = 0)

}
