% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetra-class.R
\name{deb_is_tetra}
\alias{deb_is_tetra}
\title{Test if an object is of class \code{deb_tetra}}
\usage{
deb_is_tetra(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if object is of class \code{deb_tetra} and \code{FALSE} if it is not.
}
\description{
Test if an object is of class \code{deb_tetra}.
}
\examples{
x <- deb_tetra(5, 3, 8, 2)
y <- c(5, 3, 8, 2)

deb_is_tetra(x)
deb_is_tetra(y)
}
