% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-transformations.R
\name{deb_lsd_d}
\alias{deb_lsd_d}
\title{Convert from pounds, shillings, and pence to pence}
\usage{
deb_lsd_d(lsd)
}
\arguments{
\item{lsd}{Numeric vector of length 3 or list of numeric vectors of length
3. The first position of the vector represents the pounds value or l. The
second position represents the shillings value or s. And the third
position represents the pence value or d.}
}
\value{
Returns a numeric vector of decimalized pence of length
equivalent to that of \code{l}, \code{s}, and \code{d}.
}
\description{
Convert pounds, shillings, and pence to pence or denarii.
Converting to the lowest denomination turns the non-decimal
pounds, shillings, and pence currency into a decimal currency.
}
\examples{
# Create decimalized pence
deb_lsd_d(10, 5, 8)

# The pounds, shillings, and pence do not need to be properly grouped
deb_lsd_d(l = 6, s = 25, d = 17)

# Can be used with mutate() on a data frame with
# pounds, shillings, and pence columns to create
# a column for the equivalent decimalized pence.
df <- data.frame(l = c(3, 5, 6, 2),
                 s = c(10, 18, 11, 16),
                 d = c(9, 11, 10, 5))
 dplyr::mutate(df, pence = deb_lsd_d(l, s, d))

}
