% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-decimalization-df.R
\name{deb_lsd_d_mutate}
\alias{deb_lsd_d_mutate}
\title{Conversion of pounds, shillings and pence to pence}
\usage{
deb_lsd_d_mutate(df, l = l, s = s, d = d, column_name = pence,
  bases = c(20, 12), round = 5, suffix = ".decimal")
}
\arguments{
\item{df}{A data frame that contains pounds, shillings, and pence variables.}

\item{l}{Pounds column: Unquoted name of a numeric variable corresponding
to the pounds or librae unit. Default is l.}

\item{s}{Shillings column: Unquoted name of numeric variable corresponding
to the shillings or solidi unit. Default is s.}

\item{d}{Pence column: Unquoted name of numeric variable corresponding to
the pence or denarii unit. Default is d.}

\item{column_name}{Unquoted name of the column for the decimalized
pence. Default is pence.}

\item{bases}{Numeric vector of length 2 used to specify the bases for the
shillings or s and pence or d units. Default is \code{c(20, 12)}, which
conforms to the most widely used system of 1 pound = 20 shillings and
1 shilling = 12 pence. If \code{lsd} is of class lsd or lsd_list, the bases
attribute will be used in the place of this argument.}

\item{round}{Round pence unit to specified number of decimal places.
Default is 5. Set to 0 to return pence as whole number.}

\item{suffix}{Suffix added to the column names for the pounds, shillings,
and pence columns to distinguish the new variables from the original
pounds, shillings, and pence variables if \code{replace = FALSE}. Default is
".1". Must be a character vector of length 1.}
}
\value{
Returns a data frame with a new decimalized pence variable.
}
\description{
Convert pounds, shillings, and pence variables in a data frame to a
decimalized pence variable.
}
\examples{
# Create a new decimalized pence variable
example <- data.frame(l = c(35, 10, 26, 12),
                      s = c(10, 18, 11, 16),
                      d = c(9, 11, 10, 5))
deb_lsd_d_mutate(df = example, l = l, s = s, d = d)

# Change the name of the decimalized pence variable
deb_lsd_d_mutate(df = example,
                 l = l, s = s, d = d,
                 column_name = denarii)

}
