% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-transformations.R
\name{deb_lsd_decimal}
\alias{deb_lsd_decimal}
\title{Convert from pounds, shillings and pence to pounds}
\usage{
deb_lsd_decimal(l, s, d)
}
\arguments{
\item{l}{Pounds: numeric value}

\item{s}{Shillings: numeric value}

\item{d}{Pence: numeric value}
}
\value{
Returns a single value or numeric vector of length one.
}
\description{
Convert pounds, shillings, and pence to decimalized pounds.
}
\examples{
deb_lsd_decimal(10, 5, 8)

# Can be used with mutate on a data frame with pounds, shillings, and pence
# columns to create a column for decimalized pounds.
mutate(transactions, pounds = deb_lsd_decimal(l, s, d))
}
