% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-decimalization-df.R
\name{deb_lsd_l_mutate}
\alias{deb_lsd_l_mutate}
\title{Conversion of pounds, shillings and pence to pounds}
\usage{
deb_lsd_l_mutate(df, l = l, s = s, d = d, column_name = pounds,
  bases = c(20, 12), suffix = ".decimal")
}
\arguments{
\item{df}{A data frame that contains pounds, shillings, and pence variables.}

\item{l}{Pounds column: Unquoted name of a numeric variable corresponding
to the pounds or librae unit. Default is l.}

\item{s}{Shillings column: Unquoted name of numeric variable corresponding
to the shillings or solidi unit. Default is s.}

\item{d}{Pence column: Unquoted name of numeric variable corresponding to
the pence or denarii unit. Default is d.}

\item{column_name}{Unquoted name of the column for the decimalized pounds.
Default is pounds.}

\item{bases}{Numeric vector of length 2 used to specify the bases for the
shillings or s and pence or d units. Default is \code{c(20, 12)}, which
conforms to the most widely used system of 1 pound = 20 shillings and
1 shilling = 12 pence. If \code{lsd} is of class lsd or lsd_list, the bases
attribute will be used in the place of this argument.}

\item{suffix}{Suffix added to \code{column_name} if \code{df} already contains a
variable with the same name. Default is ".decimal". Must be a character
vector of length 1.}
}
\value{
Returns a data frame with a new decimalized pounds variable.
}
\description{
Convert pounds, shillings, and pence variables in a data frame to a
decimalized pounds variable.
}
\examples{
# Create a new decimalized pounds variable
example <- data.frame(l = c(35, 10, 26, 12),
                      s = c(10, 18, 11, 16),
                      d = c(9, 11, 10, 5))
deb_lsd_l_mutate(df = example, l = l, s = s, d = d)

# Change the name of the decimalized pounds variable
deb_lsd_l_mutate(df = example,
                 l = l, s = s, d = d,
                 column_name = librae)

}
