% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-transformations.R
\name{deb_lsd_s}
\alias{deb_lsd_s}
\title{Convert from pounds, shillings and pence to shillings}
\usage{
deb_lsd_s(l, s, d)
}
\arguments{
\item{l}{Pounds: numeric value}

\item{s}{Shillings: numeric value}

\item{d}{Pence: numeric value}
}
\description{
Convert pounds, shillings, and pence to decimalized shillings.
This converts pounds, shillings, and pence to decimalized currency.
}
