% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-decimalization-df.R
\name{deb_lsd_s_mutate}
\alias{deb_lsd_s_mutate}
\title{Conversion of pounds, shillings and pence to shillings}
\usage{
deb_lsd_s_mutate(df, l = l, s = s, d = d, column_name = shillings,
  bases = c(20, 12), suffix = ".decimal")
}
\arguments{
\item{df}{A data frame that contains pounds, shillings, and pence variables.}

\item{l}{Pounds column: Unquoted name of a numeric variable corresponding
to the pounds or librae unit. Default is l.}

\item{s}{Shillings column: Unquoted name of numeric variable corresponding
to the shillings or solidi unit. Default is s.}

\item{d}{Pence column: Unquoted name of numeric variable corresponding to
the pence or denarii unit. Default is d.}

\item{column_name}{Unquoted name of the column for the decimalized
shillings. Default is shillings.}

\item{bases}{Numeric vector of length 2 used to specify the bases for
the shillings or s and pence or d units in \code{lsd}. Default is \code{c(20, 12)},
which conforms to the most widely used system of 1 pound = 20 shillings
and 1 shilling = 12 pence. This argument makes it possible to use
alternate bases for the shillings (s) and pence (d) units.}

\item{suffix}{Suffix added to the column names for the pounds, shillings,
and pence columns to distinguish the new variables from the original
pounds, shillings, and pence variables if \code{replace = FALSE}. Default is
".1". Must be a character vector of length 1.}
}
\value{
Returns a data frame with a new decimalized shillings variable.
}
\description{
Convert pounds, shillings, and pence variables in a data frame to a
decimalized shillings variable.
}
\examples{
# Create a new decimalized shillings variable
example <- data.frame(l = c(35, 10, 26, 12),
                      s = c(10, 18, 11, 16),
                      d = c(9, 11, 10, 5))
deb_lsd_s_mutate(df = example, l = l, s = s, d = d)

# Change the name of the decimalized shillings variable
deb_lsd_s_mutate(df = example,
                 l = l, s = s, d = d,
                 column_name = solidi)

}
