% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-list-column.R
\name{deb_lsd_spread}
\alias{deb_lsd_spread}
\title{Create separate pounds, shillings, and pence columns}
\usage{
deb_lsd_spread(df, lsd = lsd, l_column = l, s_column = s,
  d_column = d, suffix = ".1", replace = FALSE)
}
\arguments{
\item{df}{A data frame that contains an lsd list column.}

\item{lsd}{lsd list column: Unquoted name of an lsd list column of pounds,
shillings, and pence values with a bases attribute.}

\item{l_column}{An unquoted name for the pounds column created by the
function. Default is l.}

\item{s_column}{An unquoted name for the shillings column created by
the function. Default is s.}

\item{d_column}{An unquoted name for the pence column created by the
function. Default is d.}

\item{suffix}{Suffix added to \code{l_column}, \code{s_column}, or \code{d_column} if \code{df}
already contains variables with the same names. Default is ".1". Must be a
character vector of length 1.}

\item{replace}{Logical (default \code{FALSE}): when \code{TRUE} the original lsd list
column — \code{lsd} — will be removed.}
}
\value{
Returns a data frame with a three new variables representing pounds,
shillings, and pence with equivalent values to \code{lsd}.
}
\description{
Create separate pounds, shillings, and pence columns from an lsd list column
in a data frame.
}
\details{
The newly created pounds, shillings, and pence variables will lose their
connection to the lsd bases attribute in the lsd list column.

To return the separate l, s, and d variables to an lsd list column use
\code{\link[=deb_lsd_gather]{deb_lsd_gather()}}.
}
\examples{
library(tibble)

# Tibble with an lsd list column
example <- tibble(accounts = c(1, 2, 1, 2),
                  lsd = deb_lsd(l = c(3, 5, 6, 2),
                                s = c(10, 18, 11, 16),
                                d = c(9, 11, 10, 5)))

# Create separate l, s, and d variables
deb_lsd_spread(df = example,
               lsd = lsd)


# Can replace original lsd list column and choose the
# names for the pounds, shillings, and pence variables
deb_lsd_spread(df = example,
               lsd = lsd,
               l_column = pounds,
               s_column = shillings,
               d_column = pence,
               replace = TRUE)

}
