% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-arithmetic.R
\name{deb_multiply}
\alias{deb_multiply}
\title{Multiplication of pounds, shillings, and pence}
\usage{
deb_multiply(x, l, s, d, round = 3, vector = FALSE)
}
\arguments{
\item{x}{Multiplier. The value by which the given pounds, shillings, and
pence will be multiplied. Should be a numeric vector of length 1.}

\item{l}{Pounds: numeric vector of the same length as \code{s} and \code{d}.}

\item{s}{Shillings: numeric vector of the same length as \code{l} and \code{d}.}

\item{d}{Pence: numeric vector of the same length as \code{l} and \code{s}.}

\item{round}{Round pence to specified number of decimal places.
Default is 3. Set to 0 to return pence as whole numbers.}

\item{vector}{Logical (default \code{FALSE}): when \code{FALSE} the output will be
a tibble. When \code{TRUE} the output will be a named numeric vector or list
of named numeric vectors if the length of \code{l}, \code{s}, and \code{d} is greater
than 1.}
}
\value{
Returns either a tibble with columns for the pounds, shillings, and
pence values labeled as l, s, and d or a named numeric vector with values
for pounds, shillings, and pence. The number of rows in the resulting
tibble will be equal to the length of the input vectors. If the length of
\code{l}, \code{s}, and \code{d} is greater than 1 and \code{vector = TRUE}, the result will
be a list of named vectors of length equal to the input vectors.
}
\description{
Multiply pounds, shillings, and pence by a given multiplier (\code{x}).
The result is normalized to correct values based on 12 pence in a
shilling and 20 shillings in a pound. This is a wrapper around
\code{\link[=deb_normalize]{deb_normalize()}}.
}
\details{
This function uses the nomenclature of
\href{https://en.wikipedia.org/wiki/£sd}{l, s, and d} to refer to pounds,
shillings, and pence. This derives from the Latin terms for librae,
solidi, and denarii. One solidus was equivalent to 12 denarii, and
240 denarii coins were made from on libra of silver. The nomenclature
and values of 12 denarii to 1 solidus and 20 solidi to 1 libra was
adopted by Charlemagne and spread throughout Europe under different names.
}
\examples{
# Multiply pounds, shillings, and pence by a multiplier
deb_multiply(x = 5, l = 5, s = 15, d = 8)
deb_multiply(x = 5, 5, 15, 8, vector = TRUE)

# The l, s, and d values do not have to be normalized
deb_multiply(x = 8, 10, 38, 65)

# Like deb_normalize(), deb_multiply() can handle negative numbers
deb_multiply(x = 5, -5, -15, -8)

# Or decimals
deb_multiply(x = 5, l = 5.625, s = 15.35, d = 8)

# l, s, and d can be vectors of length > 1
# Return a tibble with two rows
deb_multiply(x = 6,
             l = c(8, 10),
             s = c(15, 6),
             d = c(3, 9))

# Return a list with two vectors
deb_multiply(x = 6,
             l = c(8, 10),
             s = c(15, 6),
             d = c(3, 9),
             vector = TRUE)

}
\seealso{
Other lsd arithmetic functions: \code{\link{deb_divide}},
  \code{\link{deb_subtract}}
}
