% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-accounts.R
\name{deb_open}
\alias{deb_open}
\title{Calculate the current values of open accounts}
\usage{
deb_open(df, credit = credit, debit = debit, l = l, s = s, d = d,
  round = 3)
}
\arguments{
\item{df}{A data frame that contains columns with pounds, shillings,
and pence variables.}

\item{credit}{credit column: Unquoted name of the credit variable. This is
the column from which the value of the transaction is from. Default
is credit. The credit column must be of the same class as the debit
column.}

\item{debit}{debit column: Unquoted name of the debit variable. This is
the column to which the value of the transaction goes. Default
is debit. The debit column must be of the same class as the credit
column.}

\item{l}{Pounds column: Unquoted name of a numeric variable corresponding
to pounds. Default is l.}

\item{s}{Shillings column: Unquoted name of numeric variable corresponding
to shillings. Default is s.}

\item{d}{Pence column: Unquoted name of numeric variable corresponding to
pence. Default is d.}

\item{round}{round pence to specified number of decimal places.
Default is 3. Set to 0 if you want pence to always be a whole number.}
}
\value{
Returns a tibble with five columns and one row for each account
present in the credit and/or debit variables in which the current
value of pounds, shillings, and pence does not equal zero. The names
for the pounds, shillings, and pence columns correspond to the input
for \code{l}, \code{s}, \code{d}.
}
\description{
Calculate the current values in the form of pounds, shillings, and pence
for all accounts in a data frame of financial transactions and show only
those accounts that are open.
}
\details{
\code{deb_open()} is similar to \code{\link{deb_current}}, but it only
returns current values for accounts that have a value that is not balanced,
and thus not equal to zero. If accounts that should be closed appear in the
results with only a small fraction of a pence remaining, set the
\code{round} argument to a lower decimal place.

\code{deb_open()} is part of a family of functions meant to be used on
data frames that contain transactions between accounts likely contained in
an account book. The data frame should possess a similar structure to a
\href{https://www.jessesadler.com/post/network-analysis-with-r/#nodes-edges}{directed edge list}.
In this context, credit and debit variables filled with the ids of the
accounts act as the edges of the network. Following the
\href{https://en.wikipedia.org/wiki/Debits_and_credits}{nomenclature of accounting},
the credit variable represents the account from which a value goes out,
while the debit variable represents the account that receives the value.
Thus, from the credit account to the debit account.

The value for the transactions should be in the non-decimal currency of
pounds, shillings, and pence. These functions use the nomenclature of
\href{https://en.wikipedia.org/wiki/£sd}{l, s, and d} to refer to pounds,
shillings, and pence. This derives from the Latin terms for librae,
solidi, and denarii. One solidus was equivalent to 12 denarii, and
240 denarii coins were made from on libra of silver. The nomenclature
and values of 12 denarii to 1 solidus and 20 solidi to 1 libra was
adopted by Charlemagne and spread throughout Europe under different names.
}
\examples{
# Create tibble in format of a transactions data frame
trans <- tibble::tibble(credit = c("a", "b", "a", "c"),
                        debit = c("b", "a", "c", "a"),
                        l = c(10, 10, 7, 9),
                        s = c(15, 15, 11, 2),
                        d = c(6, 6, 8, 11))
# Current values of open accounts present in trans
deb_open(trans)

}
\seealso{
Other lsd account functions: \code{\link{deb_account_summary}},
  \code{\link{deb_account}}, \code{\link{deb_current}}
}
