% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{deb_ptype_check}
\alias{deb_ptype_check}
\title{Check that object is of type deb_lsd or deb_decimal}
\usage{
deb_ptype_check(x)
}
\description{
Check that object is of type deb_lsd or deb_decimal
}
\keyword{internal}
