% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactor-lsd.R
\name{deb_refactor}
\alias{deb_refactor}
\title{Refactor pounds, shillings, and pence}
\usage{
deb_refactor(l, s, d, vector = FALSE)
}
\arguments{
\item{l}{pounds}

\item{s}{shillings}

\item{d}{pence}

\item{vector}{Logical (default FALSE), when FALSE the output will
be a tibble, when TRUE the output will be a numeric vector.}
}
\value{
Returns either a tibble with one row of values and columns for the
pounds, shillings, and pence values labeled as l, s, and d or a named
numeric vector with values for pounds, shillings, and pence.
}
\description{
Refactor pounds, shillings, and pence to correct values based
on 12 pence in a shilling and 20 shillings in a pound.
}
\details{
This function uses the nomenclature of l, s, and d to refer to pounds,
shillings, and pence. This derives from the Latin terms for librae,
solidi, and denarii. One solidus was equivalent to 12 denarii, and
240 denarii coins were made from on libra of silver. The nomenclature
and values of 12 denarii to 1 solidus and 20 solidi to 1 libra was
adopted by Charlemagne and spread throughout Europe under different names.
}
\examples{
deb_refactor(5, 25, 22)
deb_refactor(5, 25, 22, vector = TRUE)

}
