% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-transformations.R
\name{deb_s_lsd}
\alias{deb_s_lsd}
\title{Convert from decimalized shillings to pounds, shillings, and pence}
\usage{
deb_s_lsd(s, round = 3, vector = FALSE)
}
\arguments{
\item{s}{Shillings: numeric vector of the same length as l and d.}

\item{round}{round pence to specified number of decimal places.
Default is 3. Set to 0 if you want pence to always be a whole number.}

\item{vector}{Logical (default \code{FALSE}): when \code{FALSE} the output
will be a tibble. When \code{TRUE} the output will be a named numeric
vector or list of named numeric vectors if the length of l, s, and d is
greater than 1.}
}
\value{
Returns either a tibble with one row of values and columns for the
pounds, shillings, and pence values labeled as l, s, and d or a named
numeric vector with values for pounds, shillings, and pence. If s
is negative, the pounds, shillings, and pence values will all be negative.
}
\description{
Convert decimalized shillings to the lsd system of pounds, shillings,
and pence. The function returns the value from a decimal currency to the
non-decimal currency of pounds, shillings, and pence.
This is a wrapper around \code{deb_refactor()}.
}
\examples{
# Pounds, shillings, and pence as a tibble
deb_s_lsd(s = 2635)

# Pounds, shillings, and pence as named vector
deb_s_lsd(s = 2635, vector = TRUE)

# The value for shillings can be negative
deb_d_lsd(-2635)

# Or it can have a decimal
deb_d_lsd(2635.835)

}
