% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-arithmetic.R
\name{deb_subtract}
\alias{deb_subtract}
\title{Subtraction of two values of pounds, shillings, and pence}
\usage{
deb_subtract(lsd1, lsd2, bases = c(20, 12))
}
\arguments{
\item{lsd1, lsd2}{Numeric vectors of length 3 or lists of numeric vectors of
length 3. The first position of the vector represents the pounds value or
l. The second position represents the shillings value or s. And the third
position represents the pence value or d. lsd2 is the value or values to be
subtracted from lsd1: \code{lsd1 - lsd2}.}

\item{bases}{Numeric vector of length 2 used to specify the bases for
the shillings or s and pence or d units in \code{lsd}. Default is \code{c(20, 12)},
which conforms to the most widely used system of 1 pound = 20 shillings
and 1 shilling = 12 pence. This argument makes it possible to use
alternate bases for the shillings (s) and pence (d) units.}
}
\value{
Returns either a named numeric vector of length 3 or a list of
named numeric vectors representing the values of pounds, shillings,
and pence.
}
\description{
Subtract two pounds, shillings, and pence values that are in the form of
numeric vectors or lists of numeric vectors (\code{lsd1} and \code{lsd2}). If \code{lsd1}
and \code{lsd2} are lists of different lengths, or one is a vector, the shorter
list will be recycled.
}
\examples{
# Subtract pounds, shillings, and pence
deb_subtract(lsd1 = c(56, 8, 5), lsd2 = c(19, 5, 7))

# This is equivalent to using deb_normalize
deb_normalize(lsd = c(56, 8, 5) - c(19, 5, 7))

# deb_subtract is useful when one or both of
# lsd1 and lsd2 are lists of numerical vectors

# With two lists of the same length each vector in lsd1
# will be subtracted from the corresponding vector in lsd2.
deb_subtract(lsd1 = list(c(56, 8, 5), c(27, 12, 4)),
             lsd2 = list(c(19, 5, 7), c(6, 3, 2)))

# As opposed to using deb_normalize, which will not work
\dontrun{
deb_normalize(list(c(56, 8, 5), c(27, 12, 4)) - list(c(19, 5, 7), c(6, 3, 2)))
}

# If lsd1 and lsd2 are lists of different lengths
# or one is a vector, the shorter argument is recycled.
deb_subtract(lsd1 = list(c(56, 8, 5),
                         c(27, 12, 4),
                         c(78, 14, 11)),
             lsd2 = c(6, 18, 10))

}
\seealso{
Other lsd arithmetic functions: \code{\link{deb_add_mutate}},
  \code{\link{deb_add}}, \code{\link{deb_divide_mutate}},
  \code{\link{deb_divide}},
  \code{\link{deb_multiply_mutate}},
  \code{\link{deb_multiply}},
  \code{\link{deb_subtract_mutate}},
  \code{\link{deb_sum_df}}, \code{\link{deb_sum}}
}
\concept{lsd arithmetic functions}
