% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-sum.R
\name{deb_sum}
\alias{deb_sum}
\title{Sum of pounds, shillings, and pence in lsd vectors}
\usage{
deb_sum(..., lsd_bases = c(20, 12), na.rm = FALSE)
}
\arguments{
\item{...}{Numeric vectors of length 3 and/or lists of numeric vectors of
length 3. The first position of each vector represents the pounds value
or l. The second position represents the shillings value or s. And the
third position represents the pence value or d.}

\item{lsd_bases}{Numeric vector of length 2 used to specify the bases for
the s or solidus and d or denarius values in \code{lsd} vectors. Default is
\code{c(20, 12)}, which conforms to the most widely used system of 1 libra =
20 solidi and 1 solidus = 12 denarii. This argument makes it possible to
use alternative bases for the solidus and denarius values that were also
in use.}

\item{na.rm}{logical. Passed on to \code{na.rm} argument in \code{\link[=sum]{sum()}}. Whether
missing values (NA) should be removed. Default is \code{FALSE}.}
}
\value{
Returns a named numeric vector of length 3 representing the sum of
the pounds, shillings, and pence from the lsd vectors. If the sum is
negative, the l, s, and d values will all be negative.
}
\description{
Reduces multiple numeric vectors and/or lists of numeric vectors
to a single numeric vector of length 3 representing the normalized
sum of the lsd vectors.
}
\details{
See \code{\link[=deb_add]{deb_add()}} to add an lsd vector or list of lsd vectors to the
elements of a list of lsd vectors.
}
\examples{
# Sum of multiple lsd vectors
deb_sum(c(12, 7, 9), c(5, 8, 11), c(3, 18, 5))

# Sum of a list of lsd vectors
lsd_list <- list(c(12, 7, 9), c(5, 8, 11), c(3, 18, 5))
deb_sum(lsd_list)

# Sum of a mixture of lsd vectors and list of lsd vectors
deb_sum(lsd_list, c(8, 4, 9), c(6, 19, 10))

# Can use alternative bases for solidus and denarius units
deb_sum(lsd_list, c(8, 4, 9), c(6, 19, 10), lsd_bases = c(20, 16))

}
\seealso{
Other lsd arithmetic functions: \code{\link{deb_add_mutate}},
  \code{\link{deb_add}}, \code{\link{deb_divide_mutate}},
  \code{\link{deb_divide}},
  \code{\link{deb_multiply_mutate}},
  \code{\link{deb_multiply}},
  \code{\link{deb_subtract_mutate}},
  \code{\link{deb_subtract}}, \code{\link{deb_sum_df}}
}
