% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-sum.R
\name{deb_summarise}
\alias{deb_summarise}
\title{Sum of pounds, shillings, and pence in an lsd list column}
\usage{
deb_summarise(df, ..., round = 5, na.rm = FALSE)
}
\arguments{
\item{df}{A data frame that contains pounds, shillings, and pence variables.}

\item{...}{lsd list column or columns to be summed.}

\item{round}{Round pence unit to specified number of decimal places.
Default is 5. Set to 0 to return pence as whole number.}

\item{na.rm}{Logical. Passed on to \code{na.rm} argument in \code{\link[=sum]{sum()}}. Whether
missing values (NA) should be removed. Default is \code{FALSE}.}
}
\value{
Returns a data frame with one level of grouping dropped and lsd list
columns.
}
\description{
Uses \code{\link[dplyr:summarise_at]{dplyr::summarise_at()}} to add pounds, shillings, and pence in one or
more lsd list columns in a data frame.
}
\details{
When used on a data frame without any grouping, the result will be a data
frame with a single row consisting of the lsd list columns chosen to be
summarised. When used in conjunction with \code{\link[dplyr:group_by]{dplyr::group_by()}},
\code{\link[=deb_summarise]{deb_summarise()}} will return a sum of the lsd list columns for each group.
}
\examples{
library(tibble)
library(dplyr)

# Tibble with an lsd list column
lsd_list <- deb_as_lsd(list(c(3, 10, 9),
                            c(5, 18, 11),
                            c(6, 11, 10),
                            c(2, 16, 5)),
                       bases = c(20, 12))
example <- tibble(group = c(1, 2, 1, 2),
                  lsd = lsd_list)

# Sum of the lsd list column
deb_summarise(df = example, lsd)

# Sum of groups of the lsd list column
example \%>\%
  group_by(group) \%>\%
  deb_summarise(lsd)

}
