% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debkeepr-package.R
\docType{package}
\name{debkeepr-package}
\alias{debkeepr}
\alias{debkeepr-package}
\title{debkeepr: Analysis of Non-Decimal Currencies and Double-Entry Bookkeeping}
\description{
Analysis of historical non-decimal currencies that use the
    tripartite system of pounds, shillings, and pence.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jessesadler/debkeepr}
  \item \url{https://jessesadler.github.io/debkeepr/}
  \item Report bugs at \url{https://github.com/jessesadler/debkeepr/issues}
}

}
\author{
\strong{Maintainer}: Jesse Sadler \email{jsadler@ucla.edu} (0000-0001-6081-9681)

}
\keyword{internal}
