% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debkeepr.R
\docType{package}
\name{debkeepr}
\alias{debkeepr}
\alias{debkeepr-package}
\title{debkeepr: Analysis of Accounting in Non-Decimal Currencies}
\description{
\code{deblkeepr} provides a consistent interface for analyzing non-decimal
currencies based on the tripartite system of pounds, shillings, and pence.
The package uses the nomenclature of
\href{https://en.wikipedia.org/wiki/£sd}{l, s, and d} to represent pounds,
shillings, and pence, which derives from the Latin terms
\href{https://en.wikipedia.org/wiki/French_livre}{libra},
\href{https://en.wikipedia.org/wiki/Solidus_(coin)}{solidus}, and
\href{https://en.wikipedia.org/wiki/Denarius}{denarius}. In the 8th century a
solidus came to represent 12 denarii, and 240 denarii were made from one
libra or pound of silver. The custom of counting coins in dozens (solidi)
and scores of dozens (libra) spread throughout the Carolingian Empire and
became engrained in much of Europe. However,
\href{https://en.wikipedia.org/wiki/Non-decimal_currency}{other ratios} between
libra, solidus, and denarius were also in use. The \code{lsd_bases} argument
makes it possible to specify alternative bases for the solidus and denarius
values.
}
