% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{lsd_column_check}
\alias{lsd_column_check}
\title{Check l, s, and d values}
\usage{
lsd_column_check(df, l, s, d)
}
\arguments{
\item{l}{Pounds column: Unquoted name of numeric variable.}

\item{s}{Shillings column: Unquoted name of numeric variable.}

\item{d}{Pence column: Unquoted name of numeric variable.}
}
\description{
Check l, s, and d values
}
