% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lsd_list_to_df}
\alias{lsd_list_to_df}
\title{List of lsd vectors to data frame of pounds, shillings, and pence}
\usage{
lsd_list_to_df(lsd_list)
}
\arguments{
\item{lsd_list}{A list of numeric vectors of length 3. The first position
of the vector represents the pounds value or l. The second position
represents the shillings value or s. And the third position represents
the pence value or d.}
}
\value{
Returns a tibble or \code{tbl_df} object with pounds, shillings, and
pence variables named l, s, and d respectively. The number of rows in
the tibble will be equal to the length of \code{lsd_list}.
}
\description{
Transform a list of lsd vectors into a data frame that has pounds,
shillings, and pence variables.
}
\details{
\code{lsd_list_to_df} is a helper function to make it easier to convert between
the two primary types of input objects of multiple lsd values used by
\code{debkeepr}. See \code{\link[=df_to_lsd_list]{df_to_lsd_list()}} to convert a data frame with pounds,
shillings, and pence variable to a list with lsd vectors.
}
\examples{
# Convert a list of lsd vectors to a tibble with
# pounds, shillings, and pence variables
example <- list(c(4, 34, 89), c(-9, -75, -19), c(15.85, 36.15, 56))
lsd_list_to_df(example)

}
\seealso{
Other helper functions: \code{\link{df_to_lsd_list}}
}
