% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{mathematics}
\alias{mathematics}
\alias{sum.deb_lsd}
\alias{mean.deb_lsd}
\alias{cumsum.deb_lsd}
\alias{round.deb_lsd}
\alias{signif.deb_lsd}
\alias{ceiling.deb_lsd}
\alias{floor.deb_lsd}
\alias{trunc.deb_lsd}
\title{Math group with \code{deb_lsd} objects}
\usage{
\method{sum}{deb_lsd}(..., na.rm = FALSE)

\method{mean}{deb_lsd}(x, ..., na.rm = FALSE)

\method{cumsum}{deb_lsd}(x)

\method{round}{deb_lsd}(x, digits = 0)

\method{signif}{deb_lsd}(x, digits = 6)

\method{ceiling}{deb_lsd}(x)

\method{floor}{deb_lsd}(x)

\method{trunc}{deb_lsd}(x, ...)
}
\arguments{
\item{...}{\code{deb_lsd} vectors in \code{sum()} and arguments passed on to
further methods in \code{mean()}.}

\item{na.rm}{Logical. Should missing values (including `NaN``) be removed?}

\item{x}{An object of class \code{deb_lsd}.}

\item{digits}{Integer. Indicating the number of decimal places
(\code{round()}) or significant digits (\code{signif()}) to be used.}
}
\value{
A \code{deb_lsd} vector with normalized values.
}
\description{
Math and Summary group of functions with \code{deb_lsd} objects.
Implemented functions:
\itemize{
\item \link{Summary} group: \code{sum()}, \code{any()}, and \code{all()}.
\item \link{Math} group: \code{abs()}, \code{round()}, \code{signif()}, \code{ceiling()},
\code{floor()}, \code{trunc()}, \code{cummax()}, \code{cummin()}, and \code{cumsum()}.
\item Additional generics: \code{mean()}, \code{is.nan()}, \code{is.finite()}, and
\code{is.infinite()}.
}

All other functions from the groups not currently implemented,
including \code{median()}, \code{quantile()}, and \code{summary()}.
}
\details{
\code{sum()} and \code{cumsum()} return a normalized \code{deb_lsd} values.

Round family of functions only affect the denarius (\code{d}) unit of a
\code{deb_lsd} value. All values are normalized.
}
\examples{
x <- deb_lsd(l = c(5, 8, 12),
             s = c(16, 6, 13),
             d = c(6, 11, 0))

# All values are normalized with sum and cumsum
sum(x)
cumsum(x)
mean(x)

# Round family on deb_lsd affects the denarius unit
x2 <- deb_lsd(5, 12, 5.8365)
round(x2)
round(x2, digits = 2)
signif(x2, digits = 2)
ceiling(x2)
floor(x2)
trunc(x2)

# The returned values are normalized whether
# they are positive or negative
x3 <- deb_lsd(9, 19, 11.825)
x4 <- deb_lsd(-9, -19, -11.825)
round(x3)
round(x3, digits = 1)

ceiling(x3)
floor(x4)

trunc(x3)
trunc(x4)

}
