% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debkeepr-package.R
\name{\%<-\%}
\alias{\%<-\%}
\title{Unpacking assignment}
\usage{
x \%<-\% value
}
\value{
Invisibly returns \code{value}.
}
\description{
See \code{\link[zeallot]{\%<-\%}} for more details.
}
\keyword{internal}
