% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{text}
\alias{text}
\alias{deb_text}
\alias{deb_text.default}
\alias{deb_text.deb_lsd}
\alias{deb_text.deb_decimal}
\title{Format deb_lsd and deb_decimal objects as text}
\usage{
deb_text(x, ...)

\method{deb_text}{default}(x, ...)

\method{deb_text}{deb_lsd}(x, digits = 0, currency = "£",
  s.mark = "s.", d.mark = "d.", sep = " ", big.mark = ",",
  decimal.mark = ".", suffix = "", ...)

\method{deb_text}{deb_decimal}(x, digits = 0, currency = "£",
  big.mark = ",", decimal.mark = ".", suffix = "", ...)
}
\arguments{
\item{x}{An object of class \code{deb_lsd} or \code{deb_decimal}.}

\item{...}{Arguments passed on to further methods.}

\item{digits}{Desired number of digits after the decimal mark to which to
round the numeric values. Default is 0.}

\item{currency}{Character used for the currency mark. Default is pound sign.}

\item{s.mark}{Character used following the shillings (s) unit.
Default is "s.".}

\item{d.mark}{Character used following the pence (d) unit. Default is "d.".}

\item{sep}{Character to separate pounds, shillings, and pence units.}

\item{big.mark}{Character used to mark intervals to the left of the decimal
mark. Default is "," with default \code{big.interval} of 3.}

\item{decimal.mark}{Character used for decimal mark. Default is ".".}

\item{suffix}{Character placed after the values. Default is "".}
}
\value{
A Character vector of formatted values.
}
\description{
Flexible way to nicely format deb_lsd and deb_decimal objects for
use as labels or texts.
}
\details{
\code{deb_text} is similar to \code{as.character()} in that both return a character
vector of the values of \code{deb_lsd} and \code{deb_decimal} vectors. However,
\code{as.character()} uses the normal printing method for these objects.
\code{deb_text()} provides a convenient way to nicely format \code{deb_lsd} and
\code{deb_decimal} objects for use as text or labels with options for
customization.

\code{deb_text()} uses \code{formatC()} to format the numeric values of \code{x}. Numbers
are printed in non-scientific format and trailing zeros are dropped.

All character vector arguments should be length 1.
}
\examples{
lsd <- deb_lsd(l = c(10000, 0, -10000),
               s = c(8, 0, -8),
               d = c(5.8252, 0, -5.8252))
dec <- deb_decimal(c(10000.8252, 0, -10000.8252))

deb_text(lsd)
deb_text(dec)

# Compact format for deb_lsd with suffix to distinguish currency
deb_text(lsd, s.mark = "", d.mark = "",
         sep = ".", suffix = " Flemish")

# Control the number of digits
deb_text(lsd, digits = 3)
deb_text(dec, digits = 3)

# Change big mark and decimal mark
deb_text(lsd, digits = 4, big.mark = ".", decimal.mark = ",")
deb_text(dec, digits = 4, big.mark = ".", decimal.mark = ",")

}
\seealso{
\code{\link[=formatC]{formatC()}} for further options passed to \code{...}.
}
