% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematics.R
\name{vec_math.deb_tetra}
\alias{vec_math.deb_tetra}
\title{Error message for unimplemented mathematics functions}
\usage{
\method{vec_math}{deb_tetra}(.fn, .x, ...)
}
\arguments{
\item{.fn}{A mathematical function from the base package.}

\item{.x}{A vector.}

\item{...}{Additional arguments passed to \code{.fn}.}
}
\value{
A \code{deb_tetra} vector.
}
\description{
Error message for unimplemented mathematics functions
}
