% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-accounts.R
\name{deb_account}
\alias{deb_account}
\title{Calculate credit, debit, and current value of an account}
\usage{
deb_account(df, account_id, credit = credit, debit = debit, l = l,
  s = s, d = d, lsd_bases = c(20, 12), na.rm = FALSE)
}
\arguments{
\item{df}{A data frame containing transactions between accounts that has
values represented in the form of pounds, shillings, and pence variables.
The data frame should have two variables for accounts (credit and debit)
and variables for pounds, shillings, and pence.}

\item{account_id}{The id of the account to be used to calculate the credit,
debit, and current values. The value for the account should be present in
the credit and/or debit variables.}

\item{credit}{credit column: Unquoted name of the credit variable. This is
the column from which the value of the transaction is from. Default
is credit. The credit column must be of the same class as the debit
column.}

\item{debit}{debit column: Unquoted name of the debit variable. This is
the column to which the value of the transaction goes. Default
is debit. The debit column must be of the same class as the credit
column.}

\item{l}{Pounds column: Unquoted name of a numeric variable corresponding
to pounds. Default is l.}

\item{s}{Shillings column: Unquoted name of numeric variable corresponding
to shillings. Default is s.}

\item{d}{Pence column: Unquoted name of numeric variable corresponding to
pence. Default is d.}

\item{lsd_bases}{Numeric vector of length 2 used to specify the bases for
the s or solidus and d or denarius values in \code{lsd} vectors. Default is
\code{c(20, 12)}, which conforms to the most widely used system of 1 libra =
20 solidi and 1 solidus = 12 denarii. This argument makes it possible to
use alternative bases for the solidus and denarius values that were also
in use.}

\item{na.rm}{logical. Passed on to \code{na.rm} argument in \code{\link[=sum]{sum()}}. Whether
missing values (NA) should be removed. Default is \code{FALSE}.}
}
\value{
Returns a tibble with three rows and four columns. The rows
correspond to credit, debit, and current values in the form of
pounds, shillings, and pence. The names for the pounds, shillings,
and pence columns correspond to the input for \code{l}, \code{s},
and \code{d}.
}
\description{
Calculate the total credit, debit, and the current value of a given account
in the form of pounds, shillings, and pence.
}
\details{
\code{deb_account()} is similar to \code{\link[=deb_account_summary]{deb_account_summary()}}, but
it only returns the information for one account instead of all accounts in
the credit and/or debit variables of a transactions data frame.

\code{deb_account()} is part of a family of functions meant to be used on
data frames that contain transactions between accounts likely contained in
an account book. The data frame should possess a similar structure to a
\href{https://www.jessesadler.com/post/network-analysis-with-r/#nodes-edges}{directed edge list}.
In this context, credit and debit variables filled with the ids of the
accounts act as the edges of the network. Following the
\href{https://en.wikipedia.org/wiki/Debits_and_credits}{nomenclature of accounting},
the credit variable represents the account from which a value goes out,
while the debit variable represents the account that receives the value.
Thus, from the credit account to the debit account.

\code{deb_account()} uses the nomenclature of
\href{https://en.wikipedia.org/wiki/£sd}{l, s, and d} to represent pounds,
shillings, and pence, which derives from the Latin terms
\href{https://en.wikipedia.org/wiki/French_livre}{libra},
\href{https://en.wikipedia.org/wiki/Solidus_(coin)}{solidus}, and
\href{https://en.wikipedia.org/wiki/Denarius}{denarius}. In the 8th century a
solidus came to represent 12 denarii, and 240 denarii were made from one
libra or pound of silver. The custom of counting coins in dozens (solidi)
and scores of dozens (libra) spread throughout the Carolingian Empire and
became engrained in much of Europe. However,
\href{https://en.wikipedia.org/wiki/Non-decimal_currency}{other ratios} between
libra, solidus, and denarius were also in use. The \code{lsd_bases} argument
makes it possible to specify alternative bases for the solidus and denarius
values.
}
\examples{
# Create a transactions data frame
trans <- data.frame(credit = c("a", "b", "a", "c"),
                    debit = c("b", "a", "c", "a"),
                    l = c(10, 10, 7, 9),
                    s = c(15, 15, 11, 2),
                    d = c(6, 6, 8, 11))

# Credit, debit, and current value of account "a"
deb_account(trans, account_id = "a", credit, debit, l , s, d)

}
\seealso{
Other lsd account functions: \code{\link{deb_account_summary}},
  \code{\link{deb_balance}}, \code{\link{deb_credit}},
  \code{\link{deb_current}}, \code{\link{deb_debit}},
  \code{\link{deb_open}}
}
