% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-arithmetic.R
\name{deb_add_mutate}
\alias{deb_add_mutate}
\title{Addition of pounds, shillings, and pence in a data frame}
\usage{
deb_add_mutate(df, l = l, s = s, d = d, lsd, bases = c(20, 12),
  round = 5, replace = FALSE, suffix = ".1")
}
\arguments{
\item{df}{A data frame that contains pounds, shillings, and pence variables.}

\item{l}{Pounds column: Unquoted name of a numeric variable corresponding
to the pounds or librae unit. Default is l.}

\item{s}{Shillings column: Unquoted name of numeric variable corresponding
to the shillings or solidi unit. Default is s.}

\item{d}{Pence column: Unquoted name of numeric variable corresponding to
the pence or denarii unit. Default is d.}

\item{lsd}{Numeric vector of length 3 that represents the value to be added
to the \code{l}, \code{s}, and \code{d} variables. The first position of the vector
represents the pounds value or l. The second position represents the
shillings value or s. And the third position represents the pence value
or d.}

\item{bases}{Numeric vector of length 2 used to specify the bases for the
shillings or s and pence or d units. Default is \code{c(20, 12)}, which
conforms to the most widely used system of 1 pound = 20 shillings and
1 shilling = 12 pence. If \code{lsd} is of class lsd or lsd_list, the bases
attribute will be used in the place of this argument.}

\item{round}{Round pence unit to specified number of decimal places.
Default is 5. Set to 0 to return pence as whole number.}

\item{replace}{Logical (default \code{FALSE}): when \code{TRUE} the new pounds,
shillings, and pence variables will replace the original ones.}

\item{suffix}{Suffix added to the column names for the pounds, shillings,
and pence columns representing the added values to distinguish them
from the original pounds, shillings, and pence columns. Default is ".1".
Must be a character vector of length 1.}
}
\value{
Returns a data frame with three new variables of pounds, shillings,
and pence representing the added values.
}
\description{
Uses \code{\link[dplyr:mutate]{dplyr::mutate()}} to add pounds, shillings, and pence variables by a
pounds, shillings, and pence value. The values are returned in the form of
three new variables representing the calculated pounds, shillings and pence.
}
\details{
See \code{\link[=deb_sum_df]{deb_sum_df()}} to get a sum of the lsd values in a data frame.
}
\examples{
# Add a data frame of lsd values by £5 15s. 8d
example <- data.frame(l = c(35, 10, 26, 12),
                      s = c(10, 18, 11, 16),
                      d = c(9, 11, 10, 5))
deb_add_mutate(df = example,
               l = l, s = s, d = d,
               lsd = c(5, 15, 8))

# Replace the existing pounds, shillings, and pence
# with added values
deb_add_mutate(df = example,
               l = l, s = s, d = d,
               lsd = c(5, 15, 8),
               replace = TRUE)

}
\seealso{
Other lsd arithmetic functions: \code{\link{deb_add}},
  \code{\link{deb_divide_mutate}},
  \code{\link{deb_divide}},
  \code{\link{deb_multiply_mutate}},
  \code{\link{deb_multiply}},
  \code{\link{deb_subtract_mutate}},
  \code{\link{deb_subtract}}, \code{\link{deb_sum_df}},
  \code{\link{deb_sum}}
}
\concept{lsd arithmetic functions}
