% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-conversion.R
\name{deb_convert_bases}
\alias{deb_convert_bases}
\title{Convert between different bases for shillings and pence units}
\usage{
deb_convert_bases(lsd, bases1, bases2, ratio = 1)
}
\arguments{
\item{lsd}{Numeric vector of length 3 or list of numeric vectors of length
3. The first position of the vector represents the pounds value or l. The
second position represents the shillings value or s. And the third
position represents the pence value or d.}

\item{bases1}{Numeric vector of length 2 used to specify the bases for the
shillings or s and pence or d values of \code{lsd}. \code{bases1} represents the
bases for the current \code{lsd} vector(s).}

\item{bases2}{Numeric vector of length 2 used to specify the bases for
the shillings or s and pence or d values to which \code{lsd} will be converted.}

\item{ratio}{The ratio between the two currencies that possess different
bases. This is the value by which \code{lsd} is multiplied. Numeric vector of
length 1 with the default of \code{1}.}
}
\value{
Returns either a named numeric vector of length 3 or a list of
named numeric vectors representing the values of pounds, shillings, and
pence with the bases for shillings and pence conforming to \code{bases2}.
}
\description{
Convert between pounds, shillings, and pence currencies that use different
bases for the shillings and pence units.
}
\details{
\code{deb_convert_bases()} is similar to \code{\link[=deb_exchange]{deb_exchange()}}, but the latter performs
exchange on currencies that share the same shillings and pence bases.
}
\examples{
# Conversion between pounds Flemish of 20 shillings and 12 pence
# to guilders of 20 stuivers and 16 penningen at the rate of
# 6 guilders equals £1 Flemish
deb_convert_bases(lsd = c(204, 3, 3),
                  bases1 = c(20, 12),
                  bases2 = c(20, 16),
                  ratio = 6)

# Convert from guilders to pounds Flemish
# Flip the bases argument and change the ratio
deb_convert_bases(lsd = c(1224, 19, 8),
                  bases1 = c(20, 16),
                  bases2 = c(20, 12),
                  ratio = 1/6)

# Conversion from French crowns of 60 sous and 12 deniers to
# pound sterling of 20 shillings and 12 pence at the rate of
# 72d. French crowns equals £1 sterling or 240d. sterling
deb_convert_bases(lsd = c(214, 50, 10),
                  bases1 = c(60, 12),
                  bases2 = c(20, 12),
                  ratio = 72/240)

# Base conversion can also be done in concert with deb_exchange()
# Convert from guilders to pounds sterling at the rate of 12s. Flemish
deb_convert_bases(lsd = c(1224, 19, 8),
                  bases1 = c(20, 16),
                  bases2 = c(20, 12),
                  ratio = 1/6) \%>\%
  deb_exchange(shillings_rate = 12)

# Convert a list of lsd vectors of guilders to pounds Flemish
guilders_list <- list(c(1224, 19, 8), c(101, 5, 13), c(225, 13, 15))
deb_convert_bases(lsd = guilders_list,
                  bases1 = c(20, 16),
                  bases2 = c(20, 12),
                  ratio = 1/6)

}
