% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-conversion.R
\name{deb_convert_bases_mutate}
\alias{deb_convert_bases_mutate}
\title{Convert between different bases for shillings and pence units in a data
frame}
\usage{
deb_convert_bases_mutate(df, l = l, s = s, d = d, bases1, bases2,
  ratio = 1, suffix = ".1")
}
\arguments{
\item{df}{A data frame that contains pounds, shillings, and pence variables.}

\item{l}{Pounds column: Unquoted name of a numeric variable corresponding
to pounds. Default is l.}

\item{s}{Shillings column: Unquoted name of numeric variable corresponding
to shillings. Default is s.}

\item{d}{Pence column: Unquoted name of numeric variable corresponding to
pence. Default is d.}

\item{bases1}{Numeric vector of length 2 used to specify the bases for
the s or solidus and d or denarius values in the lsd variables.
\code{bases1} represents the bases for the current lsd variables.}

\item{bases2}{Numeric vector of length 2 used to specify the bases for
the s or solidus and d or denarius values to which the lsd variables will
be converted. \code{bases2} represents the bases for the newly created
lsd variables.}

\item{ratio}{The ratio between the two currencies that possess different
bases. Numeric vector of length 1 with the default of \code{1}. This value
by which the lsd variables are multiplied to create the new values.}

\item{suffix}{Suffix added to the names of the \code{l}, \code{s}, and \code{d} columns
to create the names for the new lsd variables.}
}
\value{
Returns a data frame with three new variables of pounds, shillings,
and pence corresponding to the shillings and pence bases in \code{bases2}.
}
\description{
Uses \code{\link[dplyr:mutate]{dplyr::mutate()}} to convert pounds, shillings, and pence currencies
that use different bases for the shillings and pence variables in a data
frame. The converted values are returned in the form of three new variables
representing the calculated pounds, shillings and pence for the new bases.
}
\details{
\code{deb_convert_bases_mutate()} is similar to \code{\link[=deb_exchange_mutate]{deb_exchange_mutate()}}, but the
latter performs exchanges on currencies that share the same shillings and
pence bases.
}
\examples{
# Conversion between pounds Flemish of 20 shillings and 12 pence
# to guilders of 20 stuivers and 16 penningen at the rate of
# 6 guilders equals £1 Flemish
flemish <- data.frame(l = c(204, 82, 36),
                      s = c(3, 15, 12),
                      d = c(3, 9, 6))
deb_convert_bases_mutate(df = flemish,
                         l = l, s = s, d = d,
                         bases1 = c(20, 12),
                         bases2 = c(20, 16),
                         ratio = 6)

# Convert from guilders to pounds Flemish
# Flip the bases argument and change the ratio
guilders <- data.frame(l = c(1224, 101, 225),
                       s = c(19, 5, 13),
                       d = c(8, 13, 15))
deb_convert_bases_mutate(df = guilders,
                         l = l, s = s, d = d,
                         bases1 = c(20, 16),
                         bases2 = c(20, 12),
                         ratio = 1/6)

}
