% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-accounts.R
\name{deb_credit}
\alias{deb_credit}
\title{Calculate the total credit of accounts}
\usage{
deb_credit(df, credit = credit, l = l, s = s, d = d,
  bases = c(20, 12), round = 5, na.rm = FALSE)
}
\arguments{
\item{df}{A data frame containing transactions between accounts with
variables for the credit and debit accounts and values in the form of
pounds, shillings, and pence variables.}

\item{credit}{Credit column: Unquoted name of the credit variable. This
represents the accounts that discharge the transactional values or from
which the values derive. Default is credit. The \code{credit} column must be
of the same class as the \code{debit} column.}

\item{l}{Pounds column: Unquoted name of a numeric variable corresponding
to the pounds or librae unit. Default is l.}

\item{s}{Shillings column: Unquoted name of numeric variable corresponding
to the shillings or solidi unit. Default is s.}

\item{d}{Pence column: Unquoted name of numeric variable corresponding to
the pence or denarii unit. Default is d.}

\item{bases}{Numeric vector of length 2 used to specify the bases for the
shillings or s and pence or d units. Default is \code{c(20, 12)}, which
conforms to the most widely used system of 1 pound = 20 shillings and
1 shilling = 12 pence. If \code{lsd} is of class lsd or lsd_list, the bases
attribute will be used in the place of this argument.}

\item{round}{Round pence to specified number of decimal places. Default
is 5. Rounding of the pence value takes place throughout the function and
is not simply done at the end of the function call. This means that the
credit, debit, and current values will add together correctly. However,
the rounded values may be different than running the function without
rounding and rounding pence values at the end.}

\item{na.rm}{Logical. Passed on to \code{na.rm} argument in \code{\link[=sum]{sum()}}. Whether
missing values (NA) should be removed. Default is \code{FALSE}.}
}
\value{
Returns a tibble with four columns and one row for each account
present in the transactions data frame (\code{df}). The values represent the
total value sent by each account to other accounts within \code{df}. The names
for the pounds, shillings, and pence variables correspond to the input
for \code{l}, \code{s}, and \code{d}.
}
\description{
Calculate the total credit of accounts in a transactions data frame.
}
\details{
\code{deb_credit()} is similar to \code{\link[=deb_account_summary]{deb_account_summary()}}, but it only returns
the credit values for the accounts in \code{df}. See \code{\link[=deb_debit]{deb_debit()}} to return
the debit totals for the accounts in \code{df}.

\code{deb_credit()} is part of a family of functions meant to be used on
data frames that contain transactions between accounts in an account
book. The data frame should possess a similar structure to a
\href{https://www.jessesadler.com/post/network-analysis-with-r/#nodes-edges}{directed edge list}.
In this context, credit and debit variables contain the ids of the accounts
and act as the edges of the network. Following the
\href{https://en.wikipedia.org/wiki/Debits_and_credits}{nomenclature of accounting},
the credit variable represents the account that the transactional value is
from, while the debit variable represents the account that receives the
value. Thus, from the credit account to the debit account.
}
\examples{
# Create a transactions data frame
trans <- data.frame(credit = c("a", "b", "b", "a", "c"),
                    debit = c("b", "a", "a", "c", "a"),
                    l = c(10, 6, 4, 7, 9),
                    s = c(15, 3, 11, 11, 2),
                    d = c(6, 11, 7, 8, 11))

# Total credit of accounts present in trans
deb_credit(df = trans, credit = credit,
           l = l, s = s, d = d)

}
\seealso{
Other lsd account functions: \code{\link{deb_account_summary}},
  \code{\link{deb_account}}, \code{\link{deb_balance}},
  \code{\link{deb_current}}, \code{\link{deb_debit}},
  \code{\link{deb_open}}
}
\concept{lsd account functions}
