% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-accounts.R
\name{deb_credit}
\alias{deb_credit}
\title{Calculate the total credit of accounts}
\usage{
deb_credit(df, credit = credit, l = l, s = s, d = d, lsd_bases = c(20,
  12), na.rm = FALSE)
}
\arguments{
\item{df}{A data frame containing transactions between accounts that has
values represented in the form of pounds, shillings, and pence variables.
The data frame should have two variables for accounts (credit and debit)
and variables for pounds, shillings, and pence.}

\item{credit}{credit column: Unquoted name of the credit variable. This is
the column from which the value of the transaction is from. Default
is credit. The credit column must be of the same class as the debit
column.}

\item{l}{Pounds column: Unquoted name of a numeric variable corresponding
to pounds. Default is l.}

\item{s}{Shillings column: Unquoted name of numeric variable corresponding
to shillings. Default is s.}

\item{d}{Pence column: Unquoted name of numeric variable corresponding to
pence. Default is d.}

\item{lsd_bases}{Numeric vector of length 2 used to specify the bases for
the s or solidus and d or denarius values in the lsd variables. Default is
\code{c(20, 12)}, which conforms to the most widely used system of 1 libra =
20 solidi and 1 solidus = 12 denarii. This argument makes it possible to
use alternative bases for the solidus and denarius values that were also
in use.}

\item{na.rm}{logical. Passed on to \code{na.rm} argument in \code{\link[=sum]{sum()}}. Whether
missing values (NA) should be removed. Default is \code{FALSE}.}
}
\value{
Returns a tibble with four columns and one row for each account
present in the transactions data frame (\code{df}). The values represent the
total value sent by each account to other accounts within \code{df} in the
form of pounds, shillings, and pence. The names for the pounds,
shillings, and pence columns correspond to the input for \code{l}, \code{s},
and \code{d}.
}
\description{
Calculate the total credit of accounts in a transactions data frame (\code{df})
in the form of pounds, shillings, and pence.
}
\details{
\code{deb_credit()} is similar to \code{\link[=deb_account_summary]{deb_account_summary()}}, but it only returns
the credit values for the accounts in \code{df}. See \code{\link[=deb_debit]{deb_debit()}} to return
the debit totals for the accounts in \code{df}.

\code{deb_credit()} is part of a family of functions meant to be used on
data frames that contain transactions between accounts likely contained in
an account book. The data frame should possess a similar structure to a
\href{https://www.jessesadler.com/post/network-analysis-with-r/#nodes-edges}{directed edge list}.
In this context, credit and debit variables filled with the ids of the
accounts act as the edges of the network. Following the
\href{https://en.wikipedia.org/wiki/Debits_and_credits}{nomenclature of accounting},
the credit variable represents the account from which a value goes out,
while the debit variable represents the account that receives the value.
Thus, from the credit account to the debit account.
}
\examples{
# Create a transactions data frame
trans <- data.frame(credit = c("a", "b", "a", "c"),
                    debit = c("b", "a", "c", "a"),
                    l = c(10, 10, 7, 9),
                    s = c(15, 15, 11, 2),
                    d = c(6, 6, 8, 11))

# Total credit of accounts present in trans
deb_credit(trans, credit, l, s, d)

}
\seealso{
Other lsd account functions: \code{\link{deb_account_summary}},
  \code{\link{deb_account}}, \code{\link{deb_balance}},
  \code{\link{deb_current}}, \code{\link{deb_debit}},
  \code{\link{deb_open}}
}
