% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-transformations.R
\name{deb_d_lsd}
\alias{deb_d_lsd}
\title{Convert from pence to pounds, shillings and pence}
\usage{
deb_d_lsd(d, vector = FALSE)
}
\arguments{
\item{d}{pence: numeric value. Can be either positive or negative.}

\item{vector}{Logical (default FALSE), when FALSE the output will
be a tibble, when TRUE the output will be a numeric vector.}
}
\value{
Returns either a tibble with one row of values and columns for the
pounds, shillings, and pence values labeled as l, s, and d or a named
numeric vector with values for pounds, shillings, and pence. If d
is negative, the pounds, shillings, and pence values will all be negative.
}
\description{
Convert pence or denarii to pounds, shillings, and pence. This returns
the value from a decimal curreny to the non-decimal currency of pounds,
shillings, and pence. This is a wrapper around \code{deb_refactor()}.
}
\examples{
# Pounds, shillings, and pence as a tibble
deb_d_lsd(d = 2500)

# Pounds, shillings, and pence as named vector
deb_d_lsd(d = 2500, vector = TRUE)

# The value for pence can be negative
deb_d_lsd(-2500)

}
