% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-transformations.R
\name{deb_d_lsd}
\alias{deb_d_lsd}
\title{Convert from pence to pounds, shillings, and pence}
\usage{
deb_d_lsd(d, round = 3, vector = FALSE)
}
\arguments{
\item{d}{Pence: numeric vector of the same length as \code{l} and \code{s}.}

\item{round}{Round pence to specified number of decimal places.
Default is 3. Set to 0 to return pence as whole numbers.}

\item{vector}{Logical (default \code{FALSE}): when \code{FALSE} the output
will be a tibble. When \code{TRUE} the output will be a named numeric
vector or list of named numeric vectors if the length of \code{l},
\code{s}, \code{d} is greater than 1.}
}
\value{
Returns either a tibble with columns for the pounds, shillings, and
pence values labeled as l, s, and d or a named numeric vector with values
for pounds, shillings, and pence. If the input value is negative, the
pounds, shillings, and pence values will all be negative. The number of
rows in the resulting tibble will be equal to the length of \code{d}. If
the length of \code{d} is greater than 1 and \code{vector = TRUE}, the
result will be a list of named vectors of length equal to \code{d}.
}
\description{
Convert pence to the lsd system of pounds, shillings, and pence.
The function returns the value from a decimal currency to the
non-decimal currency of pounds, shillings, and pence.
This is a wrapper around \code{\link{deb_normalize}}.
}
\examples{
# Pounds, shillings, and pence as a tibble
deb_d_lsd(d = 2500)

# Pounds, shillings, and pence as named vector
deb_d_lsd(d = 2500, vector = TRUE)

# The value for pence can be negative
deb_d_lsd(-2500)

# d can be a vector of length > 1
# Return a tibble with three rows
deb_d_lsd(c(2500, -2500, 4575))

# Return a list with three vectors
deb_d_lsd(c(2500, -2500, 4575), vector = TRUE)

}
