% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-accounts.R
\name{deb_debit}
\alias{deb_debit}
\title{Calculate the total debit of accounts}
\usage{
deb_debit(df, debit = debit, lsd = lsd, round = 5, na.rm = FALSE)
}
\arguments{
\item{df}{A data frame containing transactions between accounts and values
in the form of an lsd list column.}

\item{debit}{Debit column: Unquoted name of the debit variable. This
represents the accounts that receive the transactional values. Default is
debit. The \code{debit} column must be of the same class as the \code{credit}
column.}

\item{lsd}{lsd list column: Unquoted name of an lsd list column of pounds,
shillings, and pence values with a bases attribute.}

\item{round}{Round pence to specified number of decimal places. Default
is 5. Rounding of the pence value takes place throughout the function and
is not simply done at the end of the function call. This means that the
credit, debit, and current values will add together correctly. However,
the rounded values may be different than running the function without
rounding and rounding pence values at the end.}

\item{na.rm}{Logical. Passed on to \code{na.rm} argument in \code{\link[=sum]{sum()}}. Whether
missing values (NA) should be removed. Default is \code{FALSE}.}
}
\value{
Returns a tibble with one row for each account in \code{df} with the
values in an lsd list column. The values represent the total value
received by each account from other accounts within \code{df}.
}
\description{
Calculate the total debit of accounts in a transactions data frame with an
lsd list column.
}
\details{
\code{deb_debit()} is similar to \code{\link[=deb_account_summary]{deb_account_summary()}}, but it only returns
the debit values for the accounts in \code{df}. See \code{\link[=deb_credit]{deb_credit()}} to return
the credit totals for the accounts in \code{df}.

\code{deb_debit()} is part of a family of functions meant to be used on
data frames that contain transactions between accounts in an account
book. The data frame should possess a similar structure to a
\href{https://www.jessesadler.com/post/network-analysis-with-r/#nodes-edges}{directed edge list}.
In this context, credit and debit variables contain the ids of the accounts
and act as the edges of the network. Following the
\href{https://en.wikipedia.org/wiki/Debits_and_credits}{nomenclature of accounting},
the credit variable represents the account that the transactional value is
from, while the debit variable represents the account that receives the
value. Thus, from the credit account to the debit account.
}
\examples{
library(tibble)

# Create a transactions data frame
lsd_list <- deb_as_lsd(list(c(10, 15, 6),
                            c(6, 3, 11),
                            c(4, 11, 7),
                            c(7, 11, 8),
                            c(9, 2, 11)),
                       bases = c(20, 12))
trans <- tibble(credit = c("a", "b", "b", "a", "c"),
                debit = c("b", "a", "a", "c", "a"),
                lsd = lsd_list)

# Total debit of accounts present in trans
deb_debit(df = trans,
           debit = debit,
           lsd = lsd)

}
\seealso{
Other lsd account functions: \code{\link{deb_account_summary}},
  \code{\link{deb_account}}, \code{\link{deb_balance}},
  \code{\link{deb_credit}}, \code{\link{deb_current}},
  \code{\link{deb_open}}
}
\concept{lsd account functions}
