% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{deb_df_to_list}
\alias{deb_df_to_list}
\title{Data frame of pounds, shillings, and pence to list of lsd vectors}
\usage{
deb_df_to_list(df, l = l, s = s, d = d)
}
\arguments{
\item{df}{A data frame that contains pounds, shillings, and pence variables.}

\item{l}{Pounds column: Unquoted name of a numeric variable corresponding
to the pounds or librae unit. Default is l.}

\item{s}{Shillings column: Unquoted name of numeric variable corresponding
to the shillings or solidi unit. Default is s.}

\item{d}{Pence column: Unquoted name of numeric variable corresponding to
the pence or denarii unit. Default is d.}
}
\value{
Returns a list of named numeric vectors representing the values of
pounds, shillings, and pence. The length of the list will be equal to the
number of rows in \code{df}. All variables in \code{df} aside from \code{l}, \code{s}, and \code{d}
will be dropped.
}
\description{
Transform a data frame that has pounds, shillings, and pence variables into
a list of lsd vectors.
}
\details{
\code{df_to_lsd_list} is a helper function to make it easier to convert between
the two primary types of input objects of multiple lsd values used by
\code{debkeepr}. See \code{\link[=deb_list_to_df]{deb_list_to_df()}} to convert a list of lsd vectors to a
data frame with pounds, shillings, and pence variables.
}
\examples{
# Convert a data frame with pounds, shillings, and pence variables
# into a list of lsd vectors
example <- data.frame(l = c(4, -9, 25),
                      s = c(14, -5, 15),
                      d = c(9, -1, 6))
deb_df_to_list(example, l = l, s = s, d = d)

# All variables aside from l, s, and d will be dropped
example2 <- data.frame(credit = c("a", "b", "a", "c"),
                       debit = c("b", "a", "c", "a"),
                       l = c(10, 10, 7, 9),
                       s = c(15, 15, 11, 2),
                       d = c(6, 6, 8, 11))
deb_df_to_list(example2, l = l, s = s, d = d)

}
\seealso{
Other helper functions: \code{\link{deb_list_to_df}}
}
\concept{helper functions}
