% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-arithmetic.R
\name{deb_divide}
\alias{deb_divide}
\title{Division of pounds, shillings, and pence}
\usage{
deb_divide(lsd, x, bases = c(20, 12), round = 5)
}
\arguments{
\item{lsd}{An lsd value. An object of class lsd or an object that can be
coerced to class lsd: a numeric vector of length 3 or a list of such
vectors.}

\item{x}{Divisor. The value by which \code{lsd} will be divided. A numeric vector
of length 1.}

\item{bases}{Numeric vector of length 2 used to specify the bases for the
shillings or s and pence or d units. Default is \code{c(20, 12)}, which
conforms to the most widely used system of 1 pound = 20 shillings and
1 shilling = 12 pence. If \code{lsd} is of class lsd, the bases attribute will
be used in the place of this argument.}

\item{round}{Round pence unit to specified number of decimal places.
Default is 5. Set to 0 to return pence as whole number.}
}
\value{
Returns an object of class lsd with a bases attribute.
}
\description{
Divide pounds, shillings, and pence by a divisor (\code{x}). The result is
normalized to standard bases for shillings and pence units according to the
\code{bases} argument. This is a wrapper around \code{\link[=deb_normalize]{deb_normalize()}}.
}
\examples{
# Divide pounds, shillings, and pence by a divisor
deb_divide(lsd = c(63, 15, 8), x = 4)

# Divide an lsd value that has alternative bases
deb_divide(lsd = c(63, 15, 8), x = 4, bases = c(20, 16))

# Dividing an object of class lsd will use the bases attribute
lsd <- deb_as_lsd(lsd = c(63, 15, 8), bases = c(20, 16))
deb_divide(lsd = lsd, x = 4)

# The l, s, and d values do not have to be normalized
deb_divide(lsd = c(109, 38, 65), x = 8)

# Like deb_normalize(), deb_divide() can handle negative numbers
deb_divide(lsd = c(-63, -15, -8), x = 4)
deb_divide(lsd = c(63, 15, 8), x = -4)

# Or decimals
deb_divide(lsd = c(63.625, 15.35, 8), x = 4)

# Use round argument to round pence unit to specific decimal place
deb_divide(lsd = c(5, 15, 8), x = 3, round = 2)

# Divide a list of lsd values by a divisor
lsd_list <- list(c(40, 5, 9), c(29, 7, 1), c(35, 6, 5))
deb_divide(lsd = lsd_list, x = 4)

# Or an lsd object with alternative bases
lsd_list2 <- deb_as_lsd(lsd = lsd_list, bases = c(20, 16))
deb_divide(lsd = lsd_list2, x = 4)

}
\seealso{
Other lsd arithmetic functions: \code{\link{deb_add}},
  \code{\link{deb_multiply}}, \code{\link{deb_subtract}},
  \code{\link{deb_sum}}
}
\concept{lsd arithmetic functions}
