% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-exchange.R
\name{deb_exchange}
\alias{deb_exchange}
\title{Exchange between pounds, shillings and pence currencies}
\usage{
deb_exchange(lsd, shillings_rate, bases = c(20, 12))
}
\arguments{
\item{lsd}{Numeric vector of length 3 or list of numeric vectors of length
3. The first position of the vector represents the pounds value or l. The
second position represents the shillings value or s. And the third
position represents the pence value or d.}

\item{shillings_rate}{The exchange rate. This follows the common practice
of calculating the exchange rate between different currencies in terms
of shillings. Thus, in terms of a given number over the base of shillings
provided in the \code{bases} argument. A numeric vector of length 1.}

\item{bases}{Numeric vector of length 2 used to specify the bases for
the shillings or s and pence or d units in \code{lsd}. Default is \code{c(20, 12)},
which conforms to the most widely used system of 1 pound = 20 shillings
and 1 shilling = 12 pence. This argument makes it possible to use
alternate bases for the shillings (s) and pence (d) units.}
}
\value{
Returns either a named numeric vector of length 3 or a list of
named numeric vectors representing the values of pounds, shillings,
and pence.
}
\description{
Exchange between pounds, shillings, and pence currencies that share the
same bases for shillings and pence given an exchange rate calculated on
the basis of shillings.
}
\details{
\code{deb_exchange()} is a wrapper around \code{\link[=deb_multiply]{deb_multiply()}}. It uses shillings
as the basis for the conversion between the currencies. If the conversion
between currencies is presented in pounds or pence, it may be easier to
use \code{\link[=deb_multiply]{deb_multiply()}}. If the bases for the shillings and pence units differ
between the two currencies use \code{\link[=deb_convert_bases]{deb_convert_bases()}}.
}
\examples{
# Exchange at the rate of 31 shillings
deb_exchange(lsd = c(850, 16, 5), shillings_rate = 31)

# If the exchange rate is in shillings and pence, you can either
# decimalize the shillings or add the pence and divide by 12 in
# the shillings_rate argument. If the decimalized shillings has
# a repeating decimal, the latter approach is preferable.

# Exchange at the rate of 31 shillings 4 pence
deb_exchange(lsd = c(850, 16, 5), shillings_rate = 31.33333)
deb_exchange(lsd = c(850, 16, 5), shillings_rate = 31 + 4/12)

# Exchange of a list of lsd vectors at a single rate
# This returns a list of named lsd values
lsd_list <- list(c(40, 5, 9), c(29, 7, 1), c(35, 6, 5))

deb_exchange(lsd = lsd_list, shillings_rate = 31)

}
