% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-interest.R
\name{deb_interest}
\alias{deb_interest}
\title{Calculation of the interest of pounds, shillings, and pence}
\usage{
deb_interest(lsd, interest = 0.0625, duration = 1,
  with_principal = TRUE, bases = c(20, 12), round = 5)
}
\arguments{
\item{lsd}{An lsd value. An object of class lsd or an object that can be
coerced to class lsd: a numeric vector of length 3 or a list of such
vectors.}

\item{interest}{Interest rate in decimal form. A numeric vector of length 1.
Default is 0.0625, or 6.25 percent, which was default interest rate in the
Low Countries at the end of the sixteenth century.}

\item{duration}{Duration over which the interest is calculated, or the
number of times that the interest should be added. A numeric vector of
length 1. Default is 1.}

\item{with_principal}{Logical (default \code{TRUE}): when \code{TRUE} the resulting
pounds, shillings and pence value(s) will include the principal as well as
the interest. When \code{FALSE} only the interest will be returned.}

\item{bases}{Numeric vector of length 2 used to specify the bases for the
shillings or s and pence or d units. Default is \code{c(20, 12)}, which
conforms to the most widely used system of 1 pound = 20 shillings and
1 shilling = 12 pence. If \code{lsd} is of class lsd, the bases attribute will
be used in the place of this argument.}

\item{round}{Round pence unit to specified number of decimal places.
Default is 5. Set to 0 to return pence as whole number.}
}
\value{
Returns an object of class lsd with a bases attribute.
}
\description{
Calculate the interest of pounds, shillings, and pence given an interest
rate and a duration. The function does not calculate compound interest.
}
\examples{
# Calculate the interest with the principal over a certain duration
# If £10.14.5 were lent over a period of 5 years at 6.25\%
deb_interest(lsd = c(10, 14, 5), duration = 5)

# If £10.14.5 were lent over a period of 5 years at 8\%
deb_interest(lsd = c(10, 14, 5), interest = 0.08, duration = 5)

# Or you can calculate only the interest
deb_interest(lsd = c(10, 14, 5), duration = 5, with_principal = FALSE)

# Interest for an lsd value with alternative bases
deb_interest(lsd = c(10, 14, 5), duration = 5, bases = c(20, 16))

# Interest of an object of class lsd will use the bases attribute
lsd <- deb_as_lsd(lsd = c(10, 14, 5), bases = c(20, 16))
deb_interest(lsd = lsd, duration = 5)

# Interest of a list of lsd values at a single rate
lsd_list <- list(c(40, 5, 9), c(29, 7, 1), c(35, 6, 5))
deb_interest(lsd = lsd_list, interest = 0.08, duration = 5)

# Or an lsd object with alternative bases
lsd_list2 <- deb_as_lsd(lsd = lsd_list, bases = c(20, 16))
deb_interest(lsd = lsd_list2, interest = 0.08, duration = 5)

}
