% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-interest.R
\name{deb_interest}
\alias{deb_interest}
\title{Calculate the interest of pounds, shillings, and pence}
\usage{
deb_interest(lsd, interest = 0.0625, duration = 1, with_principal = TRUE,
  round = 3)
}
\arguments{
\item{lsd}{Numeric vector of length 3 or list of numeric vectors of length
3. The first position of the vector represents the pounds value or l. The
second position represents the shillings value or s. And the third
position represents the pence value or d.}

\item{interest}{Interest rate in decimal form. A numeric vector
of length 1. Default is 0.0625, or 6.25 percent, which was default
interest rate in the Low Countries at the end of the sixteenth century.}

\item{duration}{Duration over which the interest is calculated, or
number of times that the interest should be added. A numeric vector
of length 1. Default is 1.}

\item{with_principal}{Logical (default \code{TRUE}: when \code{TRUE} the resulting
pounds, shillings and pence will include the principal as well as the
interest. When \code{FALSE} only the interest will be shown.}

\item{round}{Round pence to specified number of decimal places.
Default is 3. Set to 0 to return pence as whole numbers.}
}
\value{
Returns either a named numeric vector of length 3 or a list of
named numeric vectors representing the values of pounds, shillings,
and pence. If the input lsd value is negative, the l, s, and d values
will all be negative.
}
\description{
Calculate the interest of pounds, shillings, and pence
given an interest rate and a duration. The function does not
calculate compound interest.
}
\examples{
# Calculate the interest with the principal over a certain duration
# If £10.14.5 were lent over a period of 5 years at 6.25\%
deb_interest(c(10, 14, 5), duration = 5)

# Or you can calculate only the interest
deb_interest(c(10, 14, 5), duration = 5, with_principal = FALSE)

# Use the round argument to return pence with the desired accuracy
deb_interest(c(10, 14, 5), duration = 5, round = 0)

# Interest of a list of lsd vectors at a single rate
# This returns a list of named lsd values
lsd_list <- list(c(40, 5, 9), c(29, 7, 1), c(35, 6, 5))

deb_interest(lsd = lsd_list, duration = 5, rate = 0.08)

}
