% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-interest.R
\name{deb_interest}
\alias{deb_interest}
\title{Calculate the interest of pounds, shillings, and pence}
\usage{
deb_interest(l, s, d, interest = 0.0625, duration = 1,
  with_principal = TRUE, round = 3, vector = FALSE)
}
\arguments{
\item{l}{Pounds: numeric vector of the same length as \code{s} and \code{d}.}

\item{s}{Shillings: numeric vector of the same length as \code{l} and \code{d}.}

\item{d}{Pence: numeric vector of the same length as \code{l} and \code{s}.}

\item{interest}{Interest rate in decimal form. A numeric vector
of length 1. Default is 0.0625, or 6.25 percent, which was default
interest rate in the Low Countries at the end of the sixteenth century.}

\item{duration}{Duration over which the interest is calculated, or
number of times that the interest should be added. A numeric vector
of length 1. Default is 1.}

\item{with_principal}{Logical (default \code{TRUE}: when \code{TRUE} the resulting
pounds, shillings and pence will include the interest charged as well
as the principal. When \code{FALSE} only the interest will be shown.}

\item{round}{Round pence to specified number of decimal places.
Default is 3. Set to 0 to return pence as whole numbers.}

\item{vector}{Logical (default \code{FALSE}): when \code{FALSE} the output will be
a tibble. When \code{TRUE} the output will be a named numeric vector or list
of named numeric vectors if the length of \code{l}, \code{s}, and \code{d} is greater
than 1.}
}
\value{
Returns either a tibble with columns for the pounds, shillings, and
pence values labeled as l, s, and d or a named numeric vector with values
for pounds, shillings, and pence. The number of rows in the resulting
tibble will be equal to the length of the input vectors. If the length of
\code{l}, \code{s}, and \code{d} is greater than 1 and \code{vector = TRUE}, the result will
be a list of named vectors of length equal to the input vectors.
}
\description{
Calculate the interest of pounds, shillings, and pence
given an interest rate and a duration. The function does not
calculate compound interest.
}
\examples{
# Calculate the interest with the principal over a certain duration
# If £10.14.5 were lent over a period of 5 years
deb_interest(l = 10, s = 14, d = 5, duration = 5)
deb_interest(l = 10, s = 14, d = 5, duration = 5, vector = TRUE)

# Or you can calculate only the interest
deb_interest(l = 10, s = 14, d = 5, duration = 5, with_principal = FALSE)

# l, s, and d can be vectors of length > 1
# Return a tibble with two rows
deb_interest(l = c(8, 10), s = c(15, 6), d = c(4, 9))

# Return a list with two vectors
deb_interest(l = c(8, 10), s = c(15, 6), d = c(4, 9), vector = TRUE)

# Use the round argument to return pence with the desired accuracy
deb_interest(l = 10, s = 14, d = 5, duration = 5, round = 0)

}
