% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-decimalization.R
\name{deb_l_lsd}
\alias{deb_l_lsd}
\title{Conversion of decimalized pounds to pounds, shillings, and pence}
\usage{
deb_l_lsd(l, bases = c(20, 12), round = 5)
}
\arguments{
\item{l}{Decimalized pounds: Numeric vector representing pounds to be
converted into a pounds, shillings, and pence currency.}

\item{bases}{Numeric vector of length 2 used to specify the bases for
the shillings or s and pence or d units in \code{lsd}. Default is \code{c(20, 12)},
which conforms to the most widely used system of 1 pound = 20 shillings
and 1 shilling = 12 pence. This argument makes it possible to use
alternate bases for the shillings (s) and pence (d) units.}

\item{round}{Round pence to specified number of decimal places.
Default is 5. Set to 0 to return pence as whole numbers.}
}
\value{
Returns either a named numeric vector of length 3 or a list of
named numeric vectors representing the values of pounds, shillings,
and pence.
}
\description{
Convert decimalized pounds to the lsd system of pounds, shillings, and
pence.
}
\examples{
# Conversion from pounds to pounds, shillings, and pence
# only make sense with pounds that has a decimal.
deb_l_lsd(l = 8.625)

# With alternate bases for shillings and pence
deb_l_lsd(l = 8.625, bases = c(20, 16))

# The value can be negative
deb_l_lsd(l = -8.625)

# l can be a vector of length > 1
# Return a list of lsd vectors
deb_l_lsd(l = c(8.625, -8.625, 8))

}
