% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-decimalization.R
\name{deb_l_lsd}
\alias{deb_l_lsd}
\title{Convert from decimalized pounds to pounds, shillings, and pence}
\usage{
deb_l_lsd(l, lsd_bases = c(20, 12), round = 3)
}
\arguments{
\item{l}{Decimalized pounds: numeric vector representing pounds to be
converted into an lsd value.}

\item{lsd_bases}{Numeric vector of length 2 used to specify the bases for
the s or solidus and d or denarius values in \code{lsd} vectors. Default is
\code{c(20, 12)}, which conforms to the most widely used system of 1 libra =
20 solidi and 1 solidus = 12 denarii. This argument makes it possible to
use alternative bases for the solidus and denarius values that were also
in use.}

\item{round}{Round pence to specified number of decimal places.
Default is 3. Set to 0 to return pence as whole numbers.}
}
\value{
Returns either a named numeric vector of length 3 or a list of
named numeric vectors representing the values of pounds, shillings,
and pence. If the input lsd value is negative, the l, s, and d values
will all be negative.
}
\description{
Convert decimalized pounds to the lsd system of pounds, shillings,
and pence. The function returns the value from a decimal currency to the
non-decimal currency of pounds, shillings, and pence.
This is a wrapper around \code{\link[=deb_normalize]{deb_normalize()}}.
}
\examples{
# Conversion from pounds to pounds, shillings, and pence
# only make sense with decimalized pounds.
deb_l_lsd(l = 8.625)

# The value can be negative
deb_l_lsd(l = -8.625)

# l can be a vector of length > 1
# Return a list of lsd vectors
deb_l_lsd(l = c(8.625, -8.625, 8))

}
