% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-transformations.R
\name{deb_l_lsd}
\alias{deb_l_lsd}
\title{Convert from decimalized pounds to pounds, shillings, and pence}
\usage{
deb_l_lsd(l, round = 3, vector = FALSE)
}
\arguments{
\item{l}{Pounds: numeric value}

\item{round}{round pence to specified number of decimal places. Default is 3.
Set to 0 if you want pence to always be a whole number.}

\item{vector}{Logical (default FALSE), when FALSE the output will
be a tibble, when TRUE the output will be a numeric vector.}
}
\value{
Returns either a tibble with one row of values and columns for the
pounds, shillings, and pence values labeled as l, s, and d or a named
numeric vector with values for pounds, shillings, and pence. If l
is negative, the pounds, shillings, and pence values will all be negative.
}
\description{
Convert decimalized pounds to the lsd system of pounds, shillings,
and pence. The function returns the value from a decimal currency to the
non-decimal currency of pounds, shillings, and pence.
This is a wrapper around \code{deb_refactor()}.
}
\examples{
# Conversion from pounds to pounds, shillings, and pence
# only make sense with decimalized pounds.
deb_l_lsd(l = 8.625)

# Pounds, shillings, and pence as named vector
deb_l_lsd(l = 8.625, vector = TRUE)

# The value can be negative
deb_l_lsd(-8.625)

}
