% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-transformations.R
\name{deb_l_lsd}
\alias{deb_l_lsd}
\title{Convert from decimalized pounds to pounds, shillings, and pence}
\usage{
deb_l_lsd(l, round = 3, vector = FALSE)
}
\arguments{
\item{l}{Pounds: numeric vector of the same length as \code{s} and
\code{d}.}

\item{round}{Round pence to specified number of decimal places.
Default is 3. Set to 0 to return pence as whole numbers.}

\item{vector}{Logical (default \code{FALSE}): when \code{FALSE} the output
will be a tibble. When \code{TRUE} the output will be a named numeric
vector or list of named numeric vectors if the length of \code{l},
\code{s}, \code{d} is greater than 1.}
}
\value{
Returns either a tibble with columns for the pounds, shillings, and
pence values labeled as l, s, and d or a named numeric vector with values
for pounds, shillings, and pence. If the input value is negative, the
pounds, shillings, and pence values will all be negative. The number of
rows in the resulting tibble will be equal to the length of \code{l}. If the
length of \code{l} is greater than 1 and \code{vector = TRUE}, the result will
be a list of named vectors of length equal to \code{l}.
}
\description{
Convert decimalized pounds to the lsd system of pounds, shillings,
and pence. The function returns the value from a decimal currency to the
non-decimal currency of pounds, shillings, and pence.
This is a wrapper around \code{\link{deb_normalize}}.
}
\examples{
# Conversion from pounds to pounds, shillings, and pence
# only make sense with decimalized pounds.
deb_l_lsd(l = 8.625)

# Pounds, shillings, and pence as named vector
deb_l_lsd(l = 8.625, vector = TRUE)

# The value can be negative
deb_l_lsd(-8.625)

# l can be a vector of length > 1
# Return a tibble with three rows
deb_l_lsd(c(8.625, -8.625, 8))

# Return a list with three vectors
deb_l_lsd(c(8.625, -8.625, 8), vector = TRUE)

}
