% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-decimalization.R
\name{deb_lsd_d}
\alias{deb_lsd_d}
\title{Conversion from pounds, shillings, and pence to pence}
\usage{
deb_lsd_d(lsd, bases = c(20, 12), round = 5)
}
\arguments{
\item{lsd}{An lsd value. An object of class lsd or an object that can be
coerced to class lsd: a numeric vector of length 3 or a list of such
vectors.}

\item{bases}{Numeric vector of length 2 used to specify the bases for the
shillings or s and pence or d units. Default is \code{c(20, 12)}, which
conforms to the most widely used system of 1 pound = 20 shillings and
1 shilling = 12 pence. If \code{lsd} is of class lsd, the bases attribute will
be used in the place of this argument.}

\item{round}{Round pence unit to specified number of decimal places.
Default is 5. Set to 0 to return pence as whole number.}
}
\value{
Returns a numeric vector of decimalized pence.
}
\description{
Convert pounds, shillings, and pence to decimalized pence.
}
\examples{
# Create decimalized pence
deb_lsd_d(lsd = c(10, 5, 8))

# With alternative bases for shillings and pence
deb_lsd_d(lsd = c(10, 5, 8), bases = c(20, 16))

# Decimalization an object of class lsd will use the bases attribute
lsd <- deb_as_lsd(lsd = c(10, 5, 8), bases = c(20, 16))
deb_lsd_d(lsd = lsd)

# The lsd values can be negative
deb_lsd_d(lsd = c(-10, -5, -8))

# The pounds, shillings, and pence do not need to be normalized
deb_lsd_d(lsd = c(6, 25, 17))

# Decimalize multiple lsd values with a list of lsd values
lsd_list <- list(c(40, 5, 9), c(29, 7, 1), c(35, 6, 5))
deb_lsd_d(lsd = lsd_list)

# Or an lsd object with alternative bases
lsd_list2 <- deb_as_lsd(lsd = lsd_list, bases = c(20, 16))
deb_lsd_d(lsd = lsd_list2)

}
