% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-decimalization-df.R
\name{deb_lsd_d_mutate}
\alias{deb_lsd_d_mutate}
\title{Convert from pounds, shillings and pence to pence}
\usage{
deb_lsd_d_mutate(df, l = l, s = s, d = d, column_name = denarii)
}
\arguments{
\item{df}{A data frame that contains pounds, shillings, and pence variables.}

\item{l}{Pounds column: Unquoted name of a numeric variable corresponding
to pounds. Default is l.}

\item{s}{Shillings column: Unquoted name of numeric variable corresponding
to shillings. Default is s.}

\item{d}{Pence column: Unquoted name of numeric variable corresponding to
pence. Default is d.}

\item{column_name}{Unquoted name of the column for the decimalized
pence. Default is \code{pence}.}
}
\value{
Returns a data frame with a new decimalized pence variable.
}
\description{
Convert pounds, shillings, and pence variables in a data frame to a
decimalized pence variable. Converting to the lowest denomination turns
the non-decimal pounds, shillings, and pence currency into a decimal
currency.
}
\examples{
# Create a new decimalized pence variable
example <- data.frame(l = c(35, 10, 26, 12),
                      s = c(10, 18, 11, 16),
                      d = c(9, 11, 10, 5))
deb_lsd_d_mutate(example, l, s, d)

# Change the name of the decimlaized pence variable
deb_lsd_d_mutate(example, l, s, d, column_name = pence)

}
