% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-transformations.R
\name{deb_lsd_l}
\alias{deb_lsd_l}
\title{Convert from pounds, shillings and pence to pounds}
\usage{
deb_lsd_l(l, s, d)
}
\arguments{
\item{l}{Pounds: numeric vector of the same length as \code{s} and
\code{d}.}

\item{s}{Shillings: numeric vector of the same length as \code{l}
and \code{d}.}

\item{d}{Pence: numeric vector of the same length as \code{l} and \code{s}.}
}
\value{
Returns a numeric vector of decimalized pounds of length
equivalent to that of \code{l}, \code{s}, \code{d}.
}
\description{
Convert pounds, shillings, and pence to decimalized pounds.
}
\examples{
# Create decimalized pounds
deb_lsd_l(10, 5, 8)

# The pounds, shillings, and pence do not need to be properly grouped
deb_lsd_l(l = 6, s = 25, d = 17)

# l, s, and d can be vectors of length > 1
deb_lsd_l(l = c(8, 4, 6), s = c(87, 35, 64), d = c(46, 58, 96))

# Can be used with mutate() on a data frame with
# pounds, shillings, and pence columns to create
# a column for the equivalent decimalized pounds.
df <- tibble::tibble(l = c(3, 5, 6, 2),
                     s = c(10, 18, 11, 16),
                     d = c(9, 11, 10, 5))
 dplyr::mutate(df, pounds = deb_lsd_l(l, s, d))

}
