% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-decimalization.R
\name{deb_lsd_l}
\alias{deb_lsd_l}
\title{Conversion of pounds, shillings and pence to pounds}
\usage{
deb_lsd_l(lsd, bases = c(20, 12))
}
\arguments{
\item{lsd}{An lsd value. A vector of class lsd, a list of class lsd_list, or
an object that can be coerced to these classes, namely a numeric vector of
length 3 or a list of such vectors. The first position of the vector
represents the pounds value or l. The second position represents the
shillings value or s. And the third position represents the pence value
or d.}

\item{bases}{Numeric vector of length 2 used to specify the bases for the
shillings or s and pence or d units. Default is \code{c(20, 12)}, which
conforms to the most widely used system of 1 pound = 20 shillings and
1 shilling = 12 pence. If \code{lsd} is of class lsd or lsd_list, the bases
attribute will be used in the place of this argument.}
}
\value{
Returns a numeric vector of decimalized pounds. If \code{lsd} is a
vector or object of class lsd, the returned vector will be length 1. If
\code{lsd} is a list of vectors or object of class lsd_list, the returned
vector will be the same length as \code{lsd}.
}
\description{
Convert pounds, shillings, and pence to decimalized pounds.
}
\examples{
# Create decimalized pounds
deb_lsd_l(lsd = c(10, 5, 8))

# With alternative bases for shillings and pence
deb_lsd_l(lsd = c(10, 5, 8), bases = c(20, 16))

# Decimalization an object of class lsd will use the bases attribute
lsd <- deb_as_lsd(lsd = c(10, 5, 8), bases = c(20, 16))
deb_lsd_l(lsd = lsd)

# The lsd values can be negative
deb_lsd_l(lsd = c(-10, -5, -8))

# The pounds, shillings, and pence do not need to be normalized
deb_lsd_l(lsd = c(6, 25, 17))

# Decimalize multiple lsd values with a list of lsd values
lsd_list <- list(c(40, 5, 9), c(29, 7, 1), c(35, 6, 5))
deb_lsd_l(lsd = lsd_list)

# Or an lsd_list object with alternative bases
lsd_list2 <- deb_as_lsd(lsd = lsd_list, bases = c(20, 16))
deb_lsd_l(lsd = lsd_list2)

}
