% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-transformations.R
\name{deb_lsd_l}
\alias{deb_lsd_l}
\title{Convert from pounds, shillings and pence to pounds}
\usage{
deb_lsd_l(l, s, d)
}
\arguments{
\item{l}{Pounds: numeric vector of the same length as s and d.}

\item{s}{Shillings: numeric vector of the same length as l and d.}

\item{d}{Pence: numeric vector of the same length as l and s.}
}
\value{
Returns a single value or numeric vector of length one.
}
\description{
Convert pounds, shillings, and pence to decimalized pounds.
}
\examples{
# Create decimalized pounds
deb_lsd_l(10, 5, 8)

deb_lsd_l(l = 6, s = 25, d = 17)

# Can be used with mutate() on a data frame with
# pounds, shillings, and pence columns to create
# a column for the equivalent decimalized pounds.
df <- tibble::tibble(l = c(3, 5, 6, 2),
                     s = c(10, 18, 11, 16),
                     d = c(9, 11, 10, 5))
 mutate(df, pounds = deb_lsd_l(l, s, d))

}
