% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-transformations.R
\name{deb_lsd_s}
\alias{deb_lsd_s}
\title{Convert from pounds, shillings and pence to shillings}
\usage{
deb_lsd_s(l, s, d)
}
\arguments{
\item{l}{Pounds: numeric value}

\item{s}{Shillings: numeric value}

\item{d}{Pence: numeric value}
}
\value{
Returns a single value or numeric vector of length one.
}
\description{
Convert pounds, shillings, and pence to decimalized shillings.
}
\examples{
deb_lsd_s(10, 5, 8)

# Can be used with mutate() on a data frame with
# pounds, shillings, and pence columns to create
# a column for the total decimalized shillings.
mutate(transactions, pounds = deb_lsd_s(l, s, d))
}
