% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-arithmetic.R
\name{deb_multiply}
\alias{deb_multiply}
\title{Multiplication of pounds, shillings, and pence}
\usage{
deb_multiply(lsd, x, lsd_bases = c(20, 12), round = 3)
}
\arguments{
\item{lsd}{Numeric vector of length 3 or list of numeric vectors of length
3. The first position of the vector represents the pounds value or l. The
second position represents the shillings value or s. And the third
position represents the pence value or d.}

\item{x}{Multiplier. The value by which the given pounds, shillings, and
pence will be multiplied. Should be a numeric vector of length 1.}

\item{lsd_bases}{Numeric vector of length 2 used to specify the bases for
the s or solidus and d or denarius values in \code{lsd} vectors. Default is
\code{c(20, 12)}, which conforms to the most widely used system of 1 libra =
20 solidi and 1 solidi = 12 denarii. This argument makes it possible to
use alternative bases for the solidus and denarius values that were also
in use.}

\item{round}{Round pence to specified number of decimal places.
Default is 3. Set to 0 to return pence as whole numbers.}
}
\value{
Returns either a named numeric vector of length 3 or a list of
named numeric vectors representing the values of pounds, shillings,
and pence.
}
\description{
Multiply pounds, shillings, and pence by a given multiplier (\code{x}).
The result is normalized to standard bases for shillings and pence
according to the \code{lsd_bases} argument. This is a wrapper around
\code{\link[=deb_normalize]{deb_normalize()}}.
}
\examples{
# Multiply pounds, shillings, and pence by 5
deb_multiply(lsd = c(5, 15, 8), x = 5)

# Calculate commission of 3\% of sales of £3095 17s 6d
deb_multiply(lsd = c(3095, 17, 6), x = 0.03)

# The l, s, and d values do not have to be normalized
deb_multiply(lsd = c(10, 38, 65), x = 5)

# Like deb_normalize(), deb_multiply() can handle negative numbers
deb_multiply(lsd = c(-5, -15, -8), x = 5)
deb_multiply(lsd = c(5, 15, 8), x = -5)

# Or decimals
deb_multiply(lsd = c(5.625, 15.35, 8), x = 5)

# Multiply a list of lsd vectors by a multiplier
# This returns a list of named lsd values
lsd_list <- list(c(40, 5, 9), c(29, 7, 1), c(35, 6, 5))

deb_multiply(lsd = lsd_list, x = 5)

}
\seealso{
Other lsd arithmetic functions: \code{\link{deb_divide_mutate}},
  \code{\link{deb_divide}},
  \code{\link{deb_multiply_mutate}},
  \code{\link{deb_subtract}}
}
