% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-arithmetic.R
\name{deb_multiply_mutate}
\alias{deb_multiply_mutate}
\title{Multiplication of pounds, shillings, and pence in a data frame}
\usage{
deb_multiply_mutate(df, l = l, s = s, d = d, x, round = 3,
  replace = FALSE, suffix = ".1")
}
\arguments{
\item{df}{A data frame that contains pounds, shillings, and pence variables.}

\item{l}{Pounds column: Unquoted name of a numeric variable corresponding
to pounds. Default is l.}

\item{s}{Shillings column: Unquoted name of numeric variable corresponding
to shillings. Default is s.}

\item{d}{Pence column: Unquoted name of numeric variable corresponding to
pence. Default is d.}

\item{x}{Multiplier. The value by which the given pounds, shillings, and
pence will be multiplied. Should be a numeric vector of length 1.}

\item{round}{Round pence to specified number of decimal places.
Default is 3. Set to 0 if you want pence to always be a whole number.}

\item{replace}{Logical (default \code{FALSE}): when \code{TRUE} the new pounds,
shillings, and pence variables will replace the original ones.}

\item{suffix}{Suffix added to the column names for the pounds, shillings, and
pence columns representing the multiplied values so that they are
distinguished from the original pounds, shillings, and pence columns.
Default is ".1". Should be a character vector of length 1.}
}
\value{
Returns a data frame with three new variables of pounds, shillings,
and pence representing the multiplied values.
}
\description{
Uses \code{\link[dplyr:mutate]{dplyr::mutate()}} to multiply pounds, shillings, and pence by a given
multiplier (\code{x}). The values are returned in the form of three new variables
representing the calculated pounds, shillings and pence.
}
\examples{
# Multiply a data frame of values by 5
example <- data.frame(l = c(35, 10, 26, 12),
                      s = c(10, 18, 11, 16),
                      d = c(9, 11, 10, 5))
deb_multiply_mutate(example, l, s, d, x = 5)

# Change the suffix added to the new variables
deb_multiply_mutate(example, l, s, d,
                    x = 5,
                    suffix = ".x5")

# Replace the existing pounds, shillings, and pence
# with multiplied values
deb_multiply_mutate(example, l, s, d,
                    x = 5,
                    replace = TRUE)

}
\seealso{
Other lsd arithmetic functions: \code{\link{deb_divide_mutate}},
  \code{\link{deb_divide}}, \code{\link{deb_multiply}},
  \code{\link{deb_subtract}}
}
