% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-normalize.R
\name{deb_normalize_df}
\alias{deb_normalize_df}
\title{Normalize pounds, shillings, and pence variables in a data frame}
\usage{
deb_normalize_df(df, l = l, s = s, d = d, bases = c(20, 12),
  round = 5, replace = TRUE, suffix = ".1")
}
\arguments{
\item{df}{A data frame that contains pounds, shillings, and pence variables.}

\item{l}{Pounds column: Unquoted name of a numeric variable corresponding
to the pounds or librae unit. Default is l.}

\item{s}{Shillings column: Unquoted name of numeric variable corresponding
to the shillings or solidi unit. Default is s.}

\item{d}{Pence column: Unquoted name of numeric variable corresponding to
the pence or denarii unit. Default is d.}

\item{bases}{Numeric vector of length 2 used to specify the bases for the
shillings or s and pence or d units. Default is \code{c(20, 12)}, which
conforms to the most widely used system of 1 pound = 20 shillings and
1 shilling = 12 pence. If \code{lsd} is of class lsd or lsd_list, the bases
attribute will be used in the place of this argument.}

\item{round}{Round pence unit to specified number of decimal places.
Default is 5. Set to 0 to return pence as whole number.}

\item{replace}{Logical (default \code{TRUE}): when \code{TRUE} the new pounds,
shillings, and pence variables will replace the original variables.}

\item{suffix}{Suffix added to the column names for the pounds, shillings,
and pence columns to distinguish the new variables from the original
pounds, shillings, and pence variables if \code{replace = FALSE}. Default is
".1". Must be a character vector of length 1.}
}
\value{
Returns a data frame with normalized pounds, shillings, and pence,
variables.
}
\description{
Normalize pounds, shillings, and pence variables in a data frame to standard
unit bases.
}
\details{
\code{deb_normalize_df()} uses the nomenclature of
\href{https://en.wikipedia.org/wiki/£sd}{l, s, and d} to represent pounds,
shillings, and pence units. The abbreviations derive from the Latin terms
\href{https://en.wikipedia.org/wiki/French_livre}{libra},
\href{https://en.wikipedia.org/wiki/Solidus_(coin)}{solidus}, and
\href{https://en.wikipedia.org/wiki/Denarius}{denarius}. In the 8th century a
solidus came to represent 12 denarii, and 240 denarii were made from one
libra or pound of silver. The custom of counting coins in dozens (solidi)
and scores of dozens (librae) spread throughout the Carolingian Empire and
became engrained in much of Europe. However,
\href{https://en.wikipedia.org/wiki/Non-decimal_currency}{other bases} for the
solidus and denarius units were also in use. The \code{bases} argument makes
it possible to specify alternative bases for the solidus and denarius units.
}
\examples{
# Data frame with pounds, shillings, and pence variables
example <- data.frame(l = c(35, -10, 26.725, 12),
                      s = c(50, -48, 311.805, 76),
                      d = c(89, -181, 80, 201))
# Normalize the values of pounds, shillings, and pence
deb_normalize_df(example, l = l, s = s, d = d)

# Normalize values with alternative bases for solidus and denarius
# For instance, the Dutch system of guilders, stuivers, and penningen
deb_normalize_df(example, l = l, s = s, d = d, bases = c(20, 16))

# Round argument can be used to round pence unit to
# specific decimal place and properly normalize values
deb_normalize_df(example, l = l, s = s, d = d, round = 0)

}
