% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-normalize.R
\name{deb_normalize_df}
\alias{deb_normalize_df}
\title{Normalize pounds, shillings, and pence variables in a data frame}
\usage{
deb_normalize_df(df, l = l, s = s, d = d, round = 3, replace = TRUE,
  suffix = ".1")
}
\arguments{
\item{df}{A data frame that contains separate olumns with pounds,
shillings, and pence variables.}

\item{l}{Pounds column: Unquoted name of a numeric variable corresponding
to pounds. Default is l.}

\item{s}{Shillings column: Unquoted name of numeric variable corresponding
to shillings. Default is s.}

\item{d}{Pence column: Unquoted name of numeric variable corresponding to
pence. Default is d.}

\item{round}{Round pence to specified number of decimal places.
Default is 3. Set to 0 if you want pence to always be a whole number.}

\item{replace}{Logical (default \code{TRUE}): when \code{TRUE} the new pounds,
shillings, and pence variables will replace the original ones.}

\item{suffix}{Suffix added to the column names for the pounds, shillings,
and pence columns to distinguish them from the original pounds, shillings,
and pence columns if \code{replace = FALSE}. Default is ".1". Should be a
charactern vector of length 1.}
}
\value{
Returns a data frame with normalized pounds, shillings, and pence,
variables.
}
\description{
Normalize pounds, shillings, and pence variables in a data frame to the
correct values based on 12 pence in a shilling and 20 shillings in a pound.
}
\details{
This function uses the nomenclature of
\href{https://en.wikipedia.org/wiki/£sd}{l, s, and d} to refer to pounds,
shillings, and pence. This derives from the Latin terms for librae,
solidi, and denarii. One solidus was equivalent to 12 denarii, and
240 denarii coins were made from on libra of silver. The nomenclature
and values of 12 denarii to 1 solidus and 20 solidi to 1 libra were adopted
by at least the 8th century and spread throughout Europe through the
Carolingian Empire.
}
\examples{
# Data frame with pounds, shillings, and pence variables
example <- data.frame(l = c(35, -10, 26.725, 12),
                      s = c(50, -48, 311.85, 76),
                      d = c(89, -181, 70, 205))
# Normalize lsd values
deb_normalize_df(example, l, s, d)

}
