% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-transformations.R
\name{deb_s_lsd}
\alias{deb_s_lsd}
\title{Convert from decimalized shillings to pounds, shillings, and pence}
\usage{
deb_s_lsd(s, round = 3)
}
\arguments{
\item{round}{Round pence to specified number of decimal places.
Default is 3. Set to 0 to return pence as whole numbers.}
}
\value{
Returns either a tibble with columns for the pounds, shillings, and
pence values labeled as l, s, and d or a named numeric vector with values
for pounds, shillings, and pence. If the input value is negative, the
pounds, shillings, and pence values will all be negative. The number of
rows in the resulting tibble will be equal to the length of \code{s}. If
the length of \code{s} is greater than 1 and \code{vector = TRUE``, the result will be a list of named vectors of length equal to}s`.
}
\description{
Convert decimalized shillings to the lsd system of pounds, shillings,
and pence. The function returns the value from a decimal currency to the
non-decimal currency of pounds, shillings, and pence.
This is a wrapper around \code{\link[=deb_normalize]{deb_normalize()}}.
}
\examples{
# Pounds, shillings, and pence as a tibble
deb_s_lsd(s = 2635)

# Pounds, shillings, and pence as named vector
deb_s_lsd(s = 2635, vector = TRUE)

# The value for shillings can be negative
deb_d_lsd(-2635)

# Or it can have a decimal
deb_d_lsd(2635.835)

# s can be a vector of length > 1
# Return a tibble with three rows
deb_s_lsd(c(2635, -2635, 2635.835))

# Return a list with three vectors
deb_s_lsd(c(2635, -2635, 2635.835), vector = TRUE)

}
