% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-decimalization.R
\name{deb_s_lsd}
\alias{deb_s_lsd}
\title{Conversion of decimalized shillings to pounds, shillings, and pence}
\usage{
deb_s_lsd(s, bases = c(20, 12))
}
\arguments{
\item{s}{Decimalized shillings: Numeric vector representing shillings to be
converted into a pounds, shillings, and pence currency.}

\item{bases}{Numeric vector of length 2 used to specify the bases for
the shillings or s and pence or d units in \code{lsd}. Default is \code{c(20, 12)},
which conforms to the most widely used system of 1 pound = 20 shillings
and 1 shilling = 12 pence. This argument makes it possible to use
alternate bases for the shillings (s) and pence (d) units.}
}
\value{
Returns either a named numeric vector of length 3 or a list of
named numeric vectors representing the values of pounds, shillings,
and pence.
}
\description{
Convert decimalized shillings to the lsd system of pounds, shillings,
and pence.
}
\examples{
# Convert shillings to pounds, shillings, and pence
deb_s_lsd(s = 263)

# The value for shillings can be negative
deb_s_lsd(s = -263)

# Or it can have a decimal
deb_s_lsd(s = 263.835)

# With alternate bases for shillings and pence
deb_s_lsd(s = 263.835, bases = c(20, 16))

# s can be a vector of length > 1
# Return a list of lsd vectors
deb_s_lsd(s = c(263, -263, 263.835))

}
