% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-decimalization-df.R
\name{deb_s_lsd_mutate}
\alias{deb_s_lsd_mutate}
\title{Mutate decimal shillings into pounds, shillings, and pence variables}
\usage{
deb_s_lsd_mutate(df, solidi, l_column = l, s_column = s, d_column = d,
  lsd_bases = c(20, 12), suffix = ".1")
}
\arguments{
\item{df}{A data frame that contains a column of decimalized shillings.}

\item{solidi}{Decimalized shillings column: Unquoted name of a numeric
variable corresponding to shillings. This is the variable that will be
mutated into pounds, shillings, and pence variables.}

\item{l_column}{An unquoted name for the pounds column created by the
function. Default is l.}

\item{s_column}{An unquoted name for the shillings column created by
the function. Default is s.}

\item{d_column}{An unquoted name for the pence column created by the
function. Default is d.}

\item{lsd_bases}{Numeric vector of length 2 used to specify the bases for
the s or solidus and d or denarius values in the lsd variables. Default is
\code{c(20, 12)}, which conforms to the most widely used system of 1 libra =
20 solidi and 1 solidus = 12 denarii. This argument makes it possible to
use alternative bases for the solidus and denarius values that were also
in use.}

\item{suffix}{If the data frame already contains variables with the same
names as \code{l_column}, \code{s_column}, or \code{d_column}, this suffix will be
added to the new variables to distinguish them. Default is ".1".}
}
\value{
Returns a data frame with three new variables of pounds, shillings,
and pence.
}
\description{
Uses \code{\link[dplyr:mutate]{dplyr::mutate()}} to add equivalent pounds, shillings, and pence
variables to a data frame that contains a decimalized shillings variable.
Thus, the function converts decimalized shillings into the lsd system of
pounds, shillings, and pence.
}
\examples{
# Create equivalent pounds, shillings, and pence
# variables from a decimalized shillings variable
example <- data.frame(shillings = c(166, -166, 166.5, 236.35, -354.845))

deb_s_lsd_mutate(df = example, solidi = shillings)

# You can choose the names for the created columns
example \%>\%
  deb_s_lsd_mutate(solidi = shillings,
                   l_column = librae,
                   s_column = solidi,
                   d_column = denarii)

}
