% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-sum.R
\name{deb_sum}
\alias{deb_sum}
\title{Sum of pounds, shillings, and pence values}
\usage{
deb_sum(..., bases = c(20, 12), round = 5, na.rm = FALSE)
}
\arguments{
\item{...}{lsd values. Vectors of class lsd, lists of class lsd_list, or
objects that can be coerced to these classes, namely numeric vectors of
length 3 or lists of such vectors. All lsd and lsd_list objects must have
the same bases. The first position of the vector represents the pounds
value or l. The second position represents the shillings value or s. And
the third position represents the pence value or d.}

\item{bases}{Numeric vector of length 2 used to specify the bases for the
shillings or s and pence or d units. Default is \code{c(20, 12)}, which
conforms to the most widely used system of 1 pound = 20 shillings and
1 shilling = 12 pence. If any of the lsd values are lsd or lsd_list
objects, the bases attribute will be used in the place of this argument.
All lsd and lsd_list objects must have the same bases.}

\item{round}{Round pence unit to specified number of decimal places.
Default is 5. Set to 0 to return pence as whole number.}

\item{na.rm}{Logical. Passed on to \code{na.rm} argument in \code{\link[=sum]{sum()}}. Whether
missing values (NA) should be removed. Default is \code{FALSE}.}
}
\value{
Returns an lsd object with a bases attribute.
}
\description{
Sum multiple pounds, shillings, and pence values, reducing the values to a
single pounds, shillings, and pence value in the form of an lsd object.
}
\details{
See \code{\link[=deb_add]{deb_add()}} to add an lsd value or list of lsd values to the
elements of a list of lsd values.
}
\examples{
# Sum of multiple lsd values
deb_sum(c(12, 7, 9), c(5, 8, 11), c(3, 18, 5))

# Sum of multiple lsd values with alternative bases
deb_sum(c(12, 7, 9), c(5, 8, 11), c(3, 18, 5), bases = c(20, 16))

# If one value is an lsd object, the bases attribute will be used
lsd <- deb_as_lsd(lsd = c(12, 7, 9), bases = c(20, 16))
deb_sum(lsd, c(5, 8, 11), c(3, 18, 5))

# Sum of a lsd_list object
lsd_list <- deb_as_lsd(lsd = list(c(12, 7, 9), c(5, 8, 11), c(3, 18, 5)))
deb_sum(lsd_list)

# Sum of a mixture of lsd vectors and list of lsd vectors
deb_sum(lsd_list, c(8, 4, 9), c(6, 19, 10))

# Cannot find sum of lsd and/or lsd_list
# objects that have different bases
\dontrun{
deb_sum(lsd, lsd_list)
}

}
\seealso{
Other lsd arithmetic functions: \code{\link{deb_add_mutate}},
  \code{\link{deb_add}}, \code{\link{deb_divide_mutate}},
  \code{\link{deb_divide}},
  \code{\link{deb_multiply_mutate}},
  \code{\link{deb_multiply}},
  \code{\link{deb_subtract_mutate}},
  \code{\link{deb_subtract}}, \code{\link{deb_sum_df}}
}
\concept{lsd arithmetic functions}
