% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-sum.R
\name{deb_sum}
\alias{deb_sum}
\title{Sum of pounds, shillings, and pence columns in a list of vectors}
\usage{
deb_sum(lsd_list, lsd_bases = c(20, 12), round = 3)
}
\arguments{
\item{lsd_list}{List of numeric vectors of length 3. The first position
of each vector represents the pounds value or l. The second position
represents the shillings value or s. And the third position represents
the pence value or d.}

\item{lsd_bases}{Numeric vector of length 2 used to specify the bases for
the s or solidus and d or denarius values in \code{lsd} vectors. Default is
\code{c(20, 12)}, which conforms to the most widely used system of 1 libra =
20 solidi and 1 solidus = 12 denarii. This argument makes it possible to
use alternative bases for the solidus and denarius values that were also
in use.}

\item{round}{Round pence to specified number of decimal places.
Default is 3. Set to 0 to return pence as whole numbers.}
}
\value{
Returns a named numeric vector of length 3 representing the sum of
the pounds, shillings, and pence. If the sum is negative, the l, s, and d
values will all be negative.
}
\description{
Reduces a list of numeric vectors to a single numeric vector of length 3
representing the normalized sum of the pounds, shillings, and pence in
\code{lsd_list}.
}
\examples{
# Sum of a list of lsd vectors
lsd_list <- list(c(12, 7, 9), c(5, 8, 11), c(3, 18, 5))
deb_sum(lsd_list = lsd_list)

# Can use alternative bases for solidus and denarius units
deb_sum(lsd_list = lsd_list, lsd_bases = c(20, 16))

# deb_sum cannot be used with an lsd vector
\dontrun{
deb_sum(lsd_list = c(5, 13, 8))
}

}
