% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-class.R
\name{lsd}
\alias{lsd}
\alias{deb_as_lsd}
\alias{deb_as_lsd.default}
\alias{deb_as_lsd.numeric}
\alias{deb_as_lsd.list}
\alias{deb_as_lsd.lsd}
\alias{deb_as_lsd.lsd_list}
\title{A class for pounds, shillings and pence values}
\usage{
deb_as_lsd(lsd, bases = c(20, 12), ...)

\method{deb_as_lsd}{default}(lsd, bases = c(20, 12), ...)

\method{deb_as_lsd}{numeric}(lsd, bases = c(20, 12), ...)

\method{deb_as_lsd}{list}(lsd, bases = c(20, 12), ...)

\method{deb_as_lsd}{lsd}(lsd, bases, ...)

\method{deb_as_lsd}{lsd_list}(lsd, bases, ...)
}
\arguments{
\item{lsd}{Numeric vector of length 3 or list of numeric vectors of length
3. The first position of the vector represents the pounds value or l. The
second position represents the shillings value or s. And the third
position represents the pence value or d.}

\item{bases}{Numeric vector of length 2 used to specify the bases for the
shillings or s and pence or d units. Default is \code{c(20, 12)}, which
conforms to the most widely used system of 1 pound = 20 shillings and
1 shilling = 12 pence.}

\item{...}{Arguments passed on to further methods.}
}
\value{
Returns an lsd or lsd_list object with a bases attribute.
}
\description{
Pounds, shillings, and pence values are stored as either numeric vectors
of length 3 or lists of numeric vectors of length 3 that possess a bases
attribute to record the bases for the shillings or s and pence or d units
of the values. An lsd object is a named numeric vector with a bases
attribute. An lsd_list object builds upon this. It is a list of lsd objects
that share the same bases, and it possesses a matching bases attribute
itself.
}
\details{
The lsd class and the \code{debkeepr} package use the nomenclature of
\href{https://en.wikipedia.org/wiki/£sd}{l, s, and d} to represent pounds,
shillings, and pence units. The abbreviations derive from the Latin terms
\href{https://en.wikipedia.org/wiki/French_livre}{libra},
\href{https://en.wikipedia.org/wiki/Solidus_(coin)}{solidus}, and
\href{https://en.wikipedia.org/wiki/Denarius}{denarius}. In the 8th century a
solidus came to represent 12 denarii, and 240 denarii were made from one
libra or pound of silver. The custom of counting coins in dozens (solidi)
and scores of dozens (librae) spread throughout the Carolingian Empire and
became engrained in much of Europe. However,
\href{https://en.wikipedia.org/wiki/Non-decimal_currency}{other bases} for the
solidus and denarius units were also in use. The \code{bases} attribute makes
it possible to specify alternative bases for the solidus and denarius units.
}
\examples{
# Create lsd object for £10 6s. 8d. with a numeric vector
deb_as_lsd(lsd = c(10, 6, 8), bases = c(20, 12))

# Create lsd object for the Dutch system
# of guilders, stuivers, and penningen
deb_as_lsd(lsd = c(10, 8, 14), bases = c(20, 16))

# Create lsd_list object from a list of vectors
# with the default bases of 20s. 12d.
deb_as_lsd(lsd = list(c(10, 6, 8),
                       c(8, 13, 4),
                       c(5, 8, 10)))

# A list of lsd objects must all have the same bases,
# and will use the bases attribute of the lsd objects
lsd1 <- deb_as_lsd(lsd = c(10, 8, 14), bases = c(20, 16))
lsd2 <- deb_as_lsd(lsd = c(16, 13, 2), bases = c(20, 16))
list_lsd1 <- list(lsd1, lsd2)
deb_as_lsd(lsd = list_lsd1)

# A list of lsd objects with different bases will throw and error
lsd3 <- deb_as_lsd(lsd = c(6, 11, 8), bases = c(20, 12))
list_lsd2 <- list(lsd1, lsd2, lsd3)

\dontrun{
deb_as_lsd(list_lsd2)
}

}
