% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{lsd_check}
\alias{lsd_check}
\title{Check that l, s, and d values are numeric}
\usage{
lsd_check(l, s, d)
}
\arguments{
\item{l}{Pounds: numeric value}

\item{s}{Shillings: numeric value}

\item{d}{Pence: numeric value}
}
\description{
Check that l, s, and d values are numeric
}
