% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{mathematics}
\alias{mathematics}
\alias{sum.deb_lsd}
\alias{mean.deb_lsd}
\alias{round.deb_lsd}
\alias{signif.deb_lsd}
\alias{vec_math.deb_lsd}
\title{Mathematical functions with deb_lsd objects}
\usage{
\method{sum}{deb_lsd}(..., na.rm = FALSE)

\method{mean}{deb_lsd}(x, ..., na.rm = FALSE)

\method{round}{deb_lsd}(x, digits = 0)

\method{signif}{deb_lsd}(x, digits = 6)

\method{vec_math}{deb_lsd}(.fn, .x, ...)
}
\arguments{
\item{...}{Arguments passed on to further methods.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}

\item{x}{An object of class \code{deb_lsd}.}

\item{digits}{Integer indicating the number of decimal places (round)
or significant digits (signif) to be used.}

\item{.fn}{Used internally to enable debkeepr to work with vctrs.}

\item{.x}{An object of class \code{deb_lsd}.}
}
\value{
A \code{deb_lsd} object with normalized values.
}
\description{
Mathematical functions with pounds, shillings, and pence values as
represented by the deb_lsd class.
}
\details{
\code{sum()} and \code{mean()} return a normalized \code{deb_lsd} value.

Round family of functions normalize the value and only affects the
denarius (\code{d}) unit of a \code{deb_lsd} object.
}
\examples{

x <- deb_lsd(l = c(5, 8, 12),
             s = c(16, 6, 13),
             d = c(6, 11, 0))

# Sum and mean with a deb_lsd object
# All values are normalized
sum(x)
mean(x)

# Round family on deb_lsd affects the denarius unit
x2 <- deb_lsd(5, 12, 5.8365)
round(x2)
round(x2, digits = 2)
signif(x2, digits = 2)
ceiling(x2)
floor(x2)
trunc(x2)

# The returned values are normalized whether
# they are positive or negative
x3 <- deb_lsd(9, 19, 11.825)
x4 <- deb_lsd(-9, -19, -11.825)
round(x3)
round(x3, digits = 1)

ceiling(x3)
floor(x4)

trunc(x3)
trunc(x4)

}
