% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R, R/casting.R, R/coercion.R,
%   R/comparison.R, R/debkeepr-package.R
\name{vec_math.deb_lsd}
\alias{vec_math.deb_lsd}
\alias{vec_cast.deb_lsd}
\alias{vec_cast.deb_decimal}
\alias{vec_ptype2.deb_lsd}
\alias{vec_ptype2.deb_decimal}
\alias{vec_proxy_equal.deb_lsd}
\alias{vec_proxy_compare.deb_lsd}
\alias{vctrs-compat}
\title{vctrs compatibility functions for math, casting, coercion, and comparison}
\usage{
\method{vec_math}{deb_lsd}(.fn, .x, ...)

\method{vec_cast}{deb_lsd}(x, to, ...)

\method{vec_cast}{deb_decimal}(x, to, ...)

\method{vec_ptype2}{deb_lsd}(x, y, ...)

\method{vec_ptype2}{deb_decimal}(x, y, ...)

\method{vec_proxy_equal}{deb_lsd}(x)

\method{vec_proxy_compare}{deb_lsd}(x)
}
\arguments{
\item{.fn}{Mathematical function.}

\item{.x}{Vector.}

\item{...}{Arguments passed on to further methods.}

\item{x, y}{Vectors.}

\item{to}{Type to cast to.}
}
\description{
These functions are the extensions that allow debkeepr vectors to
work with vctrs.
}
